x86, mrst: Some drivers need to known when an SCU is available
Add a notifier so that drivers can hook into SCU availability in order to
take actions post initialisation when/if the SCU becomes available.

In the ideal world we wouldn't need this and we could avoid any init
dependancies of this form, but in practice we can't do it for some cases.

Change-Id: I8504ff8fac5e8f3ac680aaf2f5334a8284af70c2
Signed-off-by: Alan Cox <alan@linux.intel.com>