ASoC: sn95031: enable the audio PLLEN bit of MSIC
Audio files plays at faster rate for 44.1khz streams. This is caused
because audio PLL of MSIC is not enabled.  Penwell provides 19.2MHz,
which is at base rate of 48Khz. PLL has to be enabled for streams which
are not at 48KHz base rate.

This patch enables the audio PLL for all the use cases which includes
voice and music use cases.

Change-Id: Iccb1b502eaa32fb963f7cf9a013a3ac8f08d7c02
Signed-off-by: Ramesh Babu K V <ramesh.babu@intel.com>