ASoC: sn95031: mfld_machine: fix jack detection and removal issues
The hardware jack detect was implemented by board designers in such a way that
it relies on the AvP headphones to ground the driver-out signal.  This was a
work around to be able to use the older style headset jack.  The headset driver
then shorts the FET gate for the plug detect circuit to operate
Offset 0x26C bit 1 should be set high (grounded) at all times until just before
the headset is turned on

This patch fixes this by grounding headset normally. Also the jack interrupt
code reliability is improved by oring new interrupt value not overwriting

Change-Id: I2c01f40def7dfe5c3012a872d24f94c9a62a93a2
Signed-off-by: Vinod Koul <vinod.koul@intel.com>