ASoC: sn95031: sst_platform: Restore MSIC Audio PLLSRC to PLLIN after the voice call
Music playback is slow when switches from Voicecall. Also, some noise can be
observed after the phone call from the speaker.

For voice call, audio driver switches the PLLSRC to PCM1BLCK. This needs to be
switched back to PLLIN to enable the switching from voice call to music
playback.

This patch modifies PLL depending on the source.

Change-Id: Id7977c4672380f9e4c3017e576ac20f4fbc76eee
Signed-off-by: Ramesh Babu K V <ramesh.babu@intel.com>
Signed-off-by: Vaibhav Agarwal <vaibhav.agarwal@intel.com>