ASoC: mfld_machine: Enabled PM Functionality Audio codec driver
During voice call and System in S3, disconnect the call.  HAL opens the
headset-dai and plays a short tone after call is disconnected.  Audio driver
responds to Headset-close(), by disabling the PLL. But DAPM for disabling the
widgets and VAUD is trigger after 5 seconds,by this time system goes to S3 and
function to invoke vaud off is not called.This will cause power dissipation.

This patch implements PM in ASoC driver which disable VAUD when system enters
S3.

Change-Id: If5383fd70f2bfb8a0c350d1eeb2a13a44be3ef1c
Signed-off-by: Jeeja <jeeja.kp@intel.com>