[PORT FROM R2] atomisp: Fix for Unload firware causing ISP Timeout
BZ: 17272

When a firmware is marked to be unloaded, it is unloaded during
the execution of WQ. WQ call be triggered for the ISP intermediate
stages well.

For a safe unloading of accelerations, the best place would
be after the FRAME DONE interrupt. Once the FRAM DONE
interrupt is received, it is guranteed that not more firmware
are executing. So moving the unlading part after the FRAME DONE
interrupt. In this way it is safer even in case of multiple
acceleration firmwares as well.

Change-Id: I371cd1fabc301fab8071b4840c0983a7db128133
Orig-Change-Id: I0760af3d1522a6cf9b8f43b4cdba6ea72e00d543
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Reviewed-on: http://android.intel.com:8080/26679
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/28053
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>