EM/THERMAL: Port Platform Thermal Driver to K3.0
BZ: 15135

This patch ports the intel_mid_thermal driver to K3.0.
This driver was already upstreamed, but without the latest
updates. So, this patch updates the code that is already available
in K3.0 and enables the driver in Kconfig. The new driver depends
on the MSIC_GPADC driver. The updated code has been tested on
Medfield PR2 platforms. Some major updates are:
* Use of GPADC driver to manipulate the ADC channels.
* Reduce the precision loss in linear interpolation
* Use ADC to Temp conversion table
* Needs IRQ initialization in intel_msic.c to load

Change-Id: I506254d12e4edf169e426ac8b02acf9281799bd0
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Reviewed-on: http://android.intel.com:8080/26761
Reviewed-by: Gugwad, SantoshX <santoshx.gugwad@intel.com>
Reviewed-by: Mitra, DebalinaX <debalinax.mitra@intel.com>
Tested-by: Gugwad, SantoshX <santoshx.gugwad@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>