Browsing is not working
BZ: 15379

As ioctl in the tty_operations struct for kernel 3.0 requires only 3
parameters, the wrong signature in n_gsm (4 parameters) leads to an IOCTL
call to the MUX which is not handled correctly.
Because of this, the RIL is not able to request the network interface
creation to the MUX via the GSMIOC_ENABLE_NET IOCTL call, so 3G data browsing is not
possible as no network interface exists.

Change-Id: Ic2ef269b5f3f161448230e6fd88ba35b3537c133
Signed-off-by: Marc Blassin <marcx.blassin@intel.com>
Reviewed-on: http://android.intel.com:8080/25804
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Tested-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>