[PORT FROM R2] atomisp: move watchdog away from workqueue
BZ: 17272

Atomisp has an embedded watchdog inside main workqueue.
As ISR is going to handle ISP/buffers, watchdog cannot rely anymore on
workqueue to handle it. This patch creates a proper interface for the
software watchdog and makes it to be handled from ISR, which is the
proper place to track if ISP is alive.

Change-Id: I90c42fb292a9bab9ad036d3dfbd1acbf39adad79
Orig-Change-Id: I551876cc49ef539e5fac871e6ca1a2336aada60d
Signed-off-by: David Cohen <david.a.cohen@intel.com>
Reviewed-on: http://android.intel.com:8080/26721
Reviewed-by: Le Thenaff, Erwan <erwan.le.thenaff@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Kruger, Jozef <jozef.kruger@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/28055
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>