mrst: process atmel_mxt224 from SFI table
BZ:16879

At present the atmel_mxt224 value is hard coded in blackbay_pr2.c
file. This value is already part of SFI table, and should make use
of this instead of using hard coded values.

changed the name of Atmel_mxt224 to mxt224 to match with the
name used in SFI table.

Change-Id: Ic86dbfc306cf28ef534bd8f6dd1e28da486af885
Signed-off-by: Hari Kanigeri <hari.k.kanigeri@intel.com>
Signed-off-by: Prabu Renganathan <prabu.renganathan@intel.com>
Reviewed-on: http://android.intel.com:8080/26863
Reviewed-by: Liu, Hong <hong.liu@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>