[PORT FROM R2]mid_pmu: provide api to set lss01 to d0i0 in atomic context
BZ: 9458

During panic we disable pre-emption but we require eMMC0
ie., LSS01 to be in D0i0 in order to store panic logs,
but pmu_pci_set_power_state might_sleep hence we cannot use
it for this purpose.

Hence we provide a exclusive api for LSS01 to be put in D0i0.

Change-Id: Ib55a9f8b61da735e76524e84d9ecd9b2a6a6493b
Signed-off-by: Illyas Mansoor <illyas.mansoor@intel.com>
Reviewed-on: http://android.intel.com:8080/30285
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>