n_gsm.c : Use of asynchronous hangup
BZ: 19302

In Linux 3.0 a new "big tty mutex" have been implemented.
This prevent tty_vhangup to be nested checking task ID (in tty_lock).
We need to launch the hangup in another workqueue which has a different task ID.
This is done by an asynchronous hangup.

flush_work is not necessary as this is done by tty layer before releasing the tty.

Change-Id: Ifdb1a4a942861e84ab948e5c978d9e493d70297a
Signed-off-by: Vincent Pillet <vincentx.pillet@intel.com>
Reviewed-on: http://android.intel.com:8080/32192
Reviewed-by: Stoltz-douchet, OlivierX <olivierx.stoltz-douchet@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>