MFLD_DV10: Update to kernel for new IFWI
BZ: 24127

This commit removes the SFI device initialization code
needed for the touchscreen. This is now handled inside
the MFLD_DV10 IFWI as is done in the PRx platform devices.

This commit also contains a temporary workaround for
the front camera GPIO (176) not being setup by the IFWI
and put into the SFI table. Once IFWI handles this like
in the PRx phone IFWI this can be removed. For now I'm
adding this along with the above BZ, but it might get its
own and I'll link dependencies of BZ's.

Change-Id: I851ad9940b56abb0e4090b820d7b484b213c5a29
Signed-off-by: Brian Wood <brian.j.wood@intel.com>
Reviewed-on: http://android.intel.com:8080/36802
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>