Audio: Configure SSP3 for CTP in Audio driver
BZ: 25850

In Medfield, SSPx is configured by SCU FW. However, in CTP platform
it is not done.

The solution can be:-
CASE1:
If audio driver should not be involved in SSP3 ALT-FUNC config then
it has be to be done as follows:
- In IFWI UMIP header, configure the GPIO pins, so that IA FW sets the
corresponding FLIS registers.
- During Audio D0i3 entry/exit, SCU FW saves and restores these values.

CASE2:
- audio driver configures SSP3 ALT-FUNC in probe.
- During Audio D0i3 entry/exit, SCU FW saves and restores these values.

CASE3:
- audio driver configures SSP3 ALT_FUNC in probe and resume.

As a workaround solution for usecase integration, CASE3 is implemented.
If there is a power impact then we need to go for CASE1 or CASE2.

Change-Id: I13f92b9559f4efe27feb9bec926ddc88946e0e60
Signed-off-by: Vaibhav Agarwal <vaibhav.agarwal@intel.com>
Reviewed-on: http://android.intel.com:8080/37685
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>