[PORT FROM R3] alarm: Adding new threshold to avoid intensive wakeup from S3
BZ: 26623

For Mobile Internet Devices, saving power is one of the main requirement.
In general, such devices supports already some power management machanism
that allows to to optimize the power consumption (such as suspend to
ram and hibernation). But using such mechanisms could be very penalizing
in term of power consumption in the case where the wakeup from deep sleep
is triggered periodically.

Actually the suspend process starts by waken up all the devices in order
to well prepare them for a deep sleep. Experimental results shows that,
depending on the devices presents on the system, the suspend increases
significantly the power if the wakeup capable alarms are configured to be
triggered before a certain threshold.

Be careful while setting this parameter: a wrong value may increase the
power consumption. So, if no full study was performed on the target
platform, please leave it at its default value.

Change-Id: I87325e0579f9eaa1e024f4c23e8065230daeeaff
Orig-Change-Id: I6ccc031f4f6991e762e74b8c92a7fa850347f0a5
Signed-off-by: Vishwesh M Rudramuni<vishwesh.m.rudramuni@intel.com>
Signed-off-by: Aymen Zayet <aymenx.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/51578
Reviewed-by: Martin, LoicX <loicx.martin@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>