PORT FROM K3.0: mmc: SDHCI: The driver should not put the device into a low-power state
It is not required (in fact it even is not recommended) that a PCI driver's
suspend() callback save the standard configuration registers of the device,
prepare it for waking up the system, or put it into a low-power state.
All of these operations can very well be taken care of by the PCI subsystem,
without the driver's participation.

Orig-Change-Id: I69184088bda601739faf1436629805034c07241f
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/23127
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Tested-by: Gross, Mark <mark.gross@intel.com>

Change-Id: I891081eff3e529559fa9c55650860c801df61fe7
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>