atomisp: fix gdc table handling
BZ: 29626, 29628

This patch fixes the incorrect freeing of the morphing table when GDC_CAC
is enabled. When it gets enabled, we need to set the current morphing table
into the CSS again, not free it. Freeing the table that the CSS was still
using caused memory corruption which luckily only resulted in the table
being rejected but could have also lead to a corrupted table being used.

Change-Id: I67177757f32c6b4f35dd176999c59bd719586d9d
Signed-off-by: Jozef Kruger <jozef.kruger@intel.com>
Reviewed-on: http://android.intel.com:8080/41714
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>