MFLD/Battery: Clean up battery status setting logic
BZ: 17854

The current code doesn't set the battery status to NOT Charging
even if the writes to MSIC register fails, it only logs the error.

This patch sets the battery status to NOT Charging if there is a
MSIC register write fail and also cleans up the battery status
setting logic.

Change-Id: Iea83c2ba2473d523d8d36c3242ca3e5307451c0f
Signed-off-by: Ramakrishna Pallala <ramakrishna.pallala@intel.com>
Reviewed-on: http://android.intel.com:8080/47166
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>