64bit: set attributes for ioapic settings for existing irqs
BZ: 36325

Found with a 64bit kernel... but also can cause issues with a 32bit kernel.
If the ioapic is negative don't set certain attributes.

(This corrects the concerns from kernel team about breaking
existing 32bit functionality as well as adds a prink for
eventual tracking down the sfi name for dev/irq if it is
negative value.)

Change-Id: I1b1fa73d545ab26317f4d120f13c30080b406568
Signed-off-by: Brian Wood <brian.j.wood@intel.com>
Reviewed-on: http://android.intel.com:8080/61902
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>