ov8830: add support for reading EEPROM data from Abico modules
BZ: 27477

Add support for reading EEPROM data from Abico modules.
This is an intermediate solution; in the long term the le24l042cs
driver should be turned into an independent driver as it is
not really dependent on the ov8830 driver. However, this
would require that user space would access directly the
subdevice nodes.

Change-Id: Iab127323bb7df78345c1ae4e4ff0179cff7092a7
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Reviewed-on: http://android.intel.com:8080/49916
Reviewed-by: Gerasimow, Timo <timo.gerasimow@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Laakso, Antti <antti.laakso@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>