[PORT FROM R3] fmradio: set timeout 10 seconds instead of 500 jiffies in FM char driver read and remove a printk on timeout
BZ: 33530 39979

Set timeout as a multiple of HZ instead of jiffies in the FM char driver read function,
since jiffies may change with the kernel.
The timeout is set to 10 seconds as is done in the GPS char driver.
Also, do not log an error on read timeout, since it is not an error.

Orig-Change-Id: Ied50adfa56f1812d27caa671e11f8bdf08c2166e
Change-Id: I34f4919bb6bf3abc948c4b0fa85653ec7c74117e
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>
Reviewed-on: http://android.intel.com:8080/50075
Reviewed-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: Peix, Fabien <fabien.peix@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>