GFX-Display: resolved DPMS off/on issue for CMI panel on CTP+.
BZ: 33332

Its DPMS off/on process is skipped by the drv_ic_inited variable, and
DPMS on couldn't take effect as its brightness wasn't restored.

NOTE: with stress test by DPMS off/on with dozens or hundreds of times,
panel content can't be updated with page flip although the TE interrupt
is still on. So maybe there is something wrong in the page flip process
and need to track the issue and more fix in advance.

Change-Id: Id08a99329f93bdefd9e18b9285366a2becf3a60c
Signed-off-by: Austin Hu <austin.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/51625
Reviewed-by: Ai, Ke <ke.ai@intel.com>
Reviewed-by: Lu, Faxing <faxing.lu@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>