Audio: Fix regression in intel_sst_driver blocking S0ix.
BZ: 41312

This audio driver patch resolves the S0ix blocking by intel_sst_driver on CTP
and Medfield.

There was a wrong pci_id check in the intel_sst_driver, which did not allow the
driver to enter the S0ix mode for CTP and medfield.This patch fixes the same.

Change-Id: I4becd390a82b20d629c4e4620316c605eff48550
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/52354
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Hibare, PramodX <pramodx.hibare@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>