EM/Charger: Fix Charger driver showing charging even when no charger present
BZ: 41637

This patch fixes the UI showing charging even when charger is not present. The
charger driver get the notification from thermal service to set the charging
current limit. Added a check to verify if the charger is present. If not then
simply return from the function without doing anything.
The charger driver showing both present=0 and online=0 even when the charger
is present is a result of charger IC going into Hi-Z mode when we disable
the charging and then reconnect the charger, the USB enumeration does not
happen and as a result both present=0 and online=0 is being shown through sysfs

Change-Id: I15387c67b19f16802d6589cbea6bf776a1107d15
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/52997
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Kannappan, R <r.kannappan@intel.com>
Reviewed-by: Prasad, Lakshmi G <lakshmi.g.prasad@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>