EM/bcove_charger: Fix bcove_init call in probe
BZ: 43362

The PMIC init function bcove_init is not called from probe function.
This was happening because the get_battery_profile function return type
was not checked properly. Also once it's invoked it was returning error
because the current and voltage lookup was giving error. This patch
fixes these two issues and initializes the PMIC from probe function

Change-Id: I7cb75b6f45cfc3b603ff99750a31e8e1f39d3f16
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Reviewed-on: http://android.intel.com:8080/53870
Reviewed-by: Shankar, AmarX <amarx.shankar@intel.com>
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Tested-by: P, Eshwar N <eshwar.n.p@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>