EM/Charger: Fix the platform not charging when put in Hi-Z mode explicitly.
BZ: 45747

This patch fixes the issue of platform not charging when putting the charger
chip into Hi-Z mode explicitly. The inital change was done to stop the battery
from charging beyond 4.2V. So, in order to avoid battery over-charging we put
the charger into low power mode by setting the Hi-Z bit REG00[7], but this
change caused some issue. The USB driver was getting invalid VBUS
interrupt and hence in turn notify the charger driver about the USB DISCONNECT
event even though the charger was connected, so the charger was going and
stopping the charging. The assumption is, the charger into Hi-Z mode is causing
the invalid VBUS interrupt.
So to fix this issue, this patch reverts the low power mode changes by removing
the part of the code which was setting charger register the Hi-Z bit.

Change-Id: Idb7abeb782ac88af621967fdcab63fcb9e5b3c86
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/55785
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Tank, Jayesh KumarX <jayesh.kumarx.tank@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>