[PORT FROM R3]wakelock: be more agressive with backoff mechanism
BZ: 46169

Some scenarios in Mobile Internet Devices such as push based
applications, may introduce unexpected storm of wakeup interrupts.

Since suspend and resume is power hungry, let's be more agressive
when to hold a wakelock if there are frequent wakeups.

The proposed configuration in this patch will make the suspend back-off
as follow:

If 2 suspend tentative happen with a residency less than 5 secs then hold
a wake lock for 30 secs.

Change-Id: I45d4f84037a92a209031dccabfa33775774374f0
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/63753
Reviewed-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>