[PORT FROM R3] Display: fix the issue of overlay crash when plug out hdmi
BZ: 45637

This is the patch to fix the overlay crash when plug out hdmi.
The disable_unused_func should not be called in detect process,
which is mainly to check hdmi connect status. It will triger hdmi
dpms to turn off the pipe timing. When overlay switches after
turning off the timing, it will cause crash issue.

The function will be called in drm_framebuffer_cleanup when release
HDMI FB buffer, which will gurantee overlay switch to MIPI completed.

Change-Id: Iecd3c03899693af1a0f00aa91392a2902cf21c92
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/56478
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>