Fix the messy or shifted Intel logo on CTP HDMI during boot up
BZ: 46461

The Intel logo shows as skewed or shifted on different TV
resolution during the boot-up time on CTP. The hardware panel
fitter has limitation to handle certain source and destination
combinations, therefore, it shows as some issues on CTP platform.

This patch fixes this by having a workaround value to tweak
the panel fit to get around the HW issue.

Change-Id: I2231be7e9532d21ce14d0661d6ca15ef7ddc2a4b
Signed-off-by: Wei Liu <wei.w.liu@intel.com>
Signed-off-by: Arun Kannan <arun.kannan@intel.com>
Reviewed-on: http://android.intel.com:8080/57885
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Lu, Faxing <faxing.lu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>