[PORT FROM R3] n_gsm.c: Avoid race condition on dlci->net
BZ: 45501

In case we have two thread, one disabling the network, the other one closing
the gsmtty, we can check the dlci->net value, then wait for the lock, and when
we catch the lock, the dlc is already destroyed.

Safer to avoid this rare race condition on dlci->net when disabling network
and closing gsmtty.

Change-Id: Iede4a15bea06fb76cbe10dd686e1ef212aa98da4
Orig-Change-Id: I58b77fd11bba3a0cbb0e01da0282b433dc6363f2
Signed-off-by: Vincent Pillet <vincentx.pillet@intel.com>
Reviewed-on: http://android.intel.com:8080/58895
Reviewed-by: Tenoutit, FaouazX <faouazx.tenoutit@intel.com>
Reviewed-by: Berat, FredericX <fredericx.berat@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>