Fix the black screen issue during quick hotplugs
BZ: 49033

During quick hotplugs, the overlay switch fails to load the programming
registers, therefore, the HW overlay gets into an error state which is
not recoverable unless deep sleep happens.

The overlay switch sequence requires the from pipe is disabled before
the pipe select bits are toggled. However, when we disable the pipe, we
did not wait long enough to ensure the pipe is fully off. The fix waits
long enough to make sure the pipe is off.

The fix also removes the drm_helper_disable_unused_functions from kernel
since the mode setting is coming from user-level. This change can reduce
the contention of locks in kernel.

HSD: 207360, 207387

Change-Id: Id08155b0a9d3520603df9580e12cc1630e3ca4ec
Signed-off-by: Wei Liu <wei.w.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/58193
Reviewed-by: Kannan, Arun <arun.kannan@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: Tu, Yun <yun.tu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>