[PORT FROM R3]: REVERTME: mm:fix kernel hang at flush_tlb_mm
BZ: 45248

when a cpu hot unpluged, its active_mm might be a shared
a mm and its cpu bit in the mm_cpumask. Later on, when a thread runing
on another cpu and using the same mm, flush_tlb_mm would send IPI
to the offline cpu and wait there forever as the cpu is offline.

Change-Id: I88134271ebb83b5baa361117f3606459b5870215
Signed-off-by: zhang, yanmin <yanmin.zhang@intel.com>
Signed-off-by: he, bo <bo.he@intel.com>
Reviewed-on: http://android.intel.com:8080/55406
Reviewed-by: Liu, ShuoX <shuox.liu@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/57042