[PORT FROM  R3] intel_soc_pmu: s0i1 governor state is promoted to LPMP3
BZ: 48472

During code refactoring the governor vs platform state selection logic
was moved to pmu driver code, and there we introduced a bug where in
when governor requests s0i1 state and platform is able to do LPMP3 we
promote s0i1 to LPMP3 this should not be allowed since it will violate
the energy break even principle and hence will cause more power drain
and if we had not done.

Fixed this by not allowing platform states > governor states.

Change-Id: Idb6242c492ccbe32eb2ce053848f76e3f502caf0
Orig-Change-Id: I161c15f837cb4f03c28dbb33cd08b4c4d115ea52
Signed-off-by: Illyas Mansoor <illyas.mansoor@intel.com>
Reviewed-on: http://android.intel.com:8080/60502
Reviewed-by: Martin, LoicX <loicx.martin@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>