[PORT FROM R3] hsi: Fix a race condition between do_dma_complete and other functions
BZ: 42546

As do_dma_complete was setting dma_ctx->ongoning->msg to NULL in a sw_lock and
deleting the node in another sw_lock, tests on ongoning->msg became invalid in
others functions. End then test if we were not cleaned-up before trying to set
it to NULL.

This patch is moving both operation under the same sw_lock to avoid such issues.

Change-Id: I08650fc4795497ece982f8ef27974a3b756c8f78
Orig-Change-Id: I2c3117990036652a21dd77204b70b298fc4c96c5
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>
Signed-off-by: Faouaz TENOUTIT <faouazx.tenoutit@intel.com>
Reviewed-on: http://android.intel.com:8080/59487
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>