EM/Charger: Fix the keyboard not working when resuming from S3
BZ: 47835

The VBUS is getting dropped to 0V when the platform is going into S3, and
when the platform resumes, there is no way for USB driver to nofity to the
charger driver what device is plugged in.
This patch has the fix to hold the wakelock as long as the battery charging
going on or the charger is supporting the OTG devices (keyboard/mouse or any)
to prevent the platform going into S3.

Change-Id: Ib6cbba8bc07db101fb6ab8e36c9159b42c81406f
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/62507
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Govindaiah, VinayX <vinayx.govindaiah@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>