[R3-JB] EM/Battery: Fix error handling in set_chrg_enable
BZ: 53333

The set_chrg_enable function returns an error when the
same values are given consecutively. This is not an
error case and the function should just return success,
when it gets the same value as that of the previous time.
This patch fixes this behavior and also corrects the
data type of some variables inside the function.

Change-Id: If4f277c3152d2b0e647153322dd05acd9a59d416
Signed-off-by: Madhukumar <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-on: http://android.intel.com:8080/63028
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: Jena, TapanX <tapanx.jena@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>