apds990x: Use large gain to improve sensitivity when ALS is enabled
BZ: 51765

Previously, if ALS sensor is enabled in very dim environment, there
may be no interrupt triggerred and the userspace can not get the
first light value until the brightness is changed dramatically. To
avoid that, we need to improve the sensitivity by setting a large
gain when ALS is enabled.

Change-Id: Ic97009891b345307242e82ca935162212ba01b8d
Signed-off-by: Leo Yan <leo.yan@intel.com>
Reviewed-on: http://android.intel.com:8080/63110
Reviewed-by: Liu, Hong <hong.liu@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>