Fix for 16:9 mode getting dropped/detected as 4:3 mode
BZ: 41346

When parsing CEA modes in edid, some video modes have 2 video identification
codes, one for 4:3 & other for 16:9 aspect ratio & the 16:9 mode is not
getting differenticated from 4:3 mode & being reported as 4:3 mode. As a result
these modes are either dropped as a duplicate or reported as 4:3.
Fixed this by changing the mode detection logic to account for aspect ratio.

Change-Id: Iae4d44a441766d453c9fcc3f484c9343220e53ea
Signed-off-by: Sanjay Rama Reddy <sanjay.rama.reddy@intel.com>
Signed-off-by: Arun Kannan <arun.kannan@intel.com>
Reviewed-on: http://android.intel.com:8080/63636
Reviewed-by: Liu, Wei W <wei.w.liu@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Jia, Lin A <lin.a.jia@intel.com>
Reviewed-by: Dai, Yu <yu.dai@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>