EM/Charger: To fix the keyboard/mouse not working when connected to CLVp phone
BZ: 53717

The charger chip bq24192i goes into Hi-Z mode whenever battery is good and
SDP charger is connected. The charger driver needs to clear the HI-Z to resume
the charger operations.
The charger driver was clearing the Hi-Z only in case a charger was plugged in
when getting notification from USB driver. Since in case of HOST type device,
charging does not need to be enabled, if charger goes in Hi-Z, it will never be
get cleared.

This patch checks the Hi-Z on every USB connect, whether the device connected is
charger or HOST and clears it.

Change-Id: Ib7b5382083bd6d17d6c439f96655a3b5e0057c3c
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/63669
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Tested-by: P, Eshwar N <eshwar.n.p@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>