EM/THERMAL: Add slope/intercept for skin temperature prediction
BZ: 53721

This patch adds slope and intercept values to predict the skin
temperature. It also refactors the code in mid_read_temp, to
make it easier to read/maintain. While we are there, rename
front skin as skin0 and back skin as skin1 to maintain naming
consistency across the driver.

Change-Id: Ia4d78263535843a388d7eddc790e0e6ab89307e7
Signed-off-by: navjotsx <navjotx.singh@intel.com>
Reviewed-on: http://android.intel.com:8080/63713
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Mistry, Pankaj <pankaj.mistry@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: P, Eshwar N <eshwar.n.p@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>