EM/Charger: To fix the battery icon abnormality in COS mode
BZ: 51924

In COS mode the battery icon showing exclaimation mark. The reason for this
issue is that ADC temperature values read from ADC was out of boundary.
This patch fixes this issue by limiting the values returned by the ADC driver
with in the specified boundary. When the value returned is greater than MAX
value, the value will be limited to MAX which will show in uevent as OVERHEAT.

This patch also has cleanup of a function which is redundant and few variables.

Change-Id: I94e6c4913913daf3afcd3af123a1a2e1365d6f14
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/64026
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Tested-by: P, Eshwar N <eshwar.n.p@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>