[PORT FROM R3]MFLD/EM: Handle invalid or fake SDP chargers
BZ: 49191

Some non standard USB mains chargers in the market violate the
USB spec and can be deteched as SDP chargers.

This patch checks for the input current limit of the charger
30 secs after it got connected to the phone hoping that the
USB enumaration had been finished by then. If the current limit
of the charger doesnt bump up to 500mA it will be consider as
invalid charger and the driver will notify the user space
by changing the charger_health to UNSPEC_FAILURE.
Signed-off-by: Santosh <santoshx.gugwad@intel.com>

Change-Id: I20bd5119a4f2d52a2617f711fff200579eb3df49
Reviewed-on: http://android.intel.com:8080/63994
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: Jena, TapanX <tapanx.jena@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: Gugwad, SantoshX <santoshx.gugwad@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>