[Revert me]: Workaround to keep display rail ON
BZ: 56395

Using the current display panel on CTP, it is not right to turn OFF VEMMC2 rail
(keeping the V_1P80_VCCAON_DISP_TCH_FB rail ON) as this is against the display
power sequence. Due to this design flaw , we are not able to turn off the display
rails completely which today consumes around 2-3 mW power during S3.

We keep VEMMC2 rail ON temporarily, until we get a proper solution from the
display/board team.

Change-Id: Ibc0ef833a461a1a9feeab171f5153c9e2e6c07ec
Signed-off-by: Kumar Mahesh <mahesh.kumar.p@intel.com>
Reviewed-on: http://android.intel.com:8080/65640
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>