power_supply: Register battery as a thermal zone
BZ: 56906

Battery and charger contribute to Thermals in most of the embedded
devices. So, it makes sense to identify them as Thermal zones in a
particular platform.

This patch registers a thermal zone if the power supply is reporting
a temperature property. The thermal zone will be used by platform's
thermal management solution.

This patch is back ported from mainline. Modified the
thermal_zone_device_register call according to definition in the
JB kernel andcheck for WARN_ON status to handle error

Change-Id: Iab6fc7e680252d982e0e0f9854df816e85a3ea81
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Reviewed-on: http://android.intel.com:8080/65775
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Singh, NavjotX <navjotx.singh@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>