[PORT FROM R3] Fix incorrect conversion macros
BZ: 55887

the percentage to current step macro is not correct. The
reverse, step to percentage is not either. The two shoudl be
the reverse of eacht other and they are not. As a result the
outcome was not always correct. This patch makes the two marco's
each other's reverse and adds an automatic calculation of the
correction factor based on the max percentage range and the max
curent step value that is allowed.

Orig-Change-Id: Ieb3e0a3b180a635140d0802551217cec7bf7fe8d
Change-Id: I937ed75b6c235c97be187af6b4305e31d0d238a7
Signed-off-by: Kaisa Taipale <kaisa.taipale@intel.com>
Reviewed-on: http://android.intel.com:8080/66733
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Kaartinen, Jukka O <jukka.o.kaartinen@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>