atomisp: Reset sw_contex.error flag before WQ is sehduled
BZ: 36881

Whenever WQ is finished executing because of error or cancel_work_sync
call, it sets sw_context.error flag to true. And this flag is reset
when WQ is started. In atomisp_dqbuf(), it checks this flag when a DQ
request comes from the user space and if this flag is set, it
returns an error.

But there can be cases where the DQ requests comes even before the
WQ is sheduled. In such cases it should not return any error. This
patch moves the reseting of this flag from WQ to the stream_on
function

Change-Id: I70beabbee9273eaa3a8ceb3d6787fdc66e003fcc
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Reviewed-on: http://android.intel.com:8080/66956
Reviewed-by: Toivonen, Tuukka <tuukka.toivonen@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>