ov8830: Embed binfactor and skip frames info to mode definition
BZ:56142

As the sensor modes are hard coded, when a mode is selected we
already know the binning factor used for that mode. By including
this information in to the sensor mode definition, we dont need to
read the register values again when required. Skip frames are used
to inform the user side application the number of frames it needs
to skip to get the best shot from the sensor. This is especially
important when we switch from preview mode to still mode as there
will different HTS and VTS settings.

Change-Id: I616962f0643a6cfd705c9cc8859a23a66267f009
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Reviewed-on: http://android.intel.com:8080/67432
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>