[BACKPORT] ALSA: hda - use LPIB for delay estimation
BZ: 138605

DMA Position in Buffer (DPIB) should be used for
ring buffer management, while LPIB register provides
information on the number of samples transfered on
the link. The difference between the two pieces of
information corresponds to hardware/DMA buffering.

This patch reports this difference in runtime->delay, and
removes the use of the COMBO mode on recent Intel hardware.

Credits to Takashi Iwai for an initial patch.

[rebased to for-next branch and replaced snd_printk() with
 snd_printdd() by tiwai]

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Conflicts:
	sound/pci/hda/hda_intel.c

Upstream commit id: 90accc58a6946e7245993da6079f88d8c29cb731

This patch doesn't get applied as is on k3.4 and hence manual
and partial backport is necessary.

Change-Id: I3b07c2101eabbe8061ca49822e7b0db2c28cde6a
Signed-off-by: Tsai,Gaggery <gaggery.tsai@intel.com>
Signed-off-by: Ramesh Babu <Ramesh.Babu@intel.com>