[PORT_FROM_R3Lex]audio: pll handling when both pcm1 and pcm2 active at same time
BZ: 33708

In medfield based designs, PCM2 is statically configured as master.
The PCM1 is configured as slave or master based on use cases.
In some use cases, both PCM1 and PCM2 will be opened simultaneously.
If PCM1 is configured as slave and PCM2 as master, then PLL should
be synced to PCM1BLCK.

During PLL configuration, check the media(PCM2) and voice(PCM1) usage count
and based on the count and configuration of voice port program the
PLL.
Consider the following use case-
1. FM has started. Voice port is configured as slave and PLL is synced to
	PCM1BCLK.
2. User does touch action. A touch tone will be played on PCM2. During PCM2
	hw_params, machine driver needs to check the existing state of PLL
	before programming PLL. If PLL is already configured to PCM1BCLK,
	then do not modify PLL.
3. FM is closed, however touch tone is still playing. Voice port is closed.
	In voice port shutdown function, reconfigure the PLL to PLLIN since
	the media is still open.
In step #3, user might notice a small glitch. Because PLL is dynamically
reconfigured while the codec widgets are ON.

Change-Id: I855b4bdb63afc0abc6f4ad494e45337c777d7dcf
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>
Signed-off-by: Namarta Kohli <namartax.kohli@intel.com>
Reviewed-on: http://android.intel.com:8080/67630
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>