gpio: langwell: add set_debounce API to enable/disable debouncing
BZ: 59368

This implements the set_debounce API of gpiolib
to enable or disable debouncing of the gpio
line.

Since we can only enable and disable debouncing
and there is no debouncing time setting on
the chip, passing 0 as debouncing time will
disable debouncing and any other value will
enable it.

Change-Id: I5365a9e1914b5d7a8577104b548dbba654db4ca4
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Reviewed-on: http://android.intel.com:8080/67880
Reviewed-by: Li, Ning <ning.li@intel.com>
Reviewed-by: Yang, Bin <bin.yang@intel.com>
Reviewed-by: Cohen, David A <david.a.cohen@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>