gfx: Using the marco CONFIG_MDFD_GL3 to fix wrong GL3 island operation
BZ: 61031

In current code, the marco CONFIG_MDFD_GL3 is used to control the GL3
island enabling/disabling, but there are two places codes which do GL3
power on/off actions without the control of CONFIG_MDFD_GL3, it is related
with video operations.

So this patch fix it to avoid unneccessary system issue(fabric error....).

Change-Id: I46e419288590555b1f81dc0003a48b254e2b271e
Signed-off-by: liu chuansheng <chuansheng.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/69349
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: Shi, Yang A <yang.a.shi@intel.com>
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>