[PORT FROM R4-ENZO] atomisp2: Produce SOF events to user space
BZ: 61242

Get SOF interrupt from the ISP and provide V4L2_EVENT_FRAME_SYNC events to
the user space through the subdev node.  The frame_sequence field is not yet
set.

A part of this patch went in unintentionally in the threaded IRQ patches.
This is the rest.

The Merrifield CSS does not support currently SOF events, thus the ugly

Change-Id: If9635c84890574cb87b8ea9e2fb07adf92c1847a
Orig-Change-Id: I65ce954d46293c5c3dcb6549fc359e328376050b
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/71717
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>