Graphics-Display: Fix for ESD thread not resuming after sleep
BZ: 60293

It appears the ESD detection thread sometimes does not continue
if you wake the phone up from sleep.  It appears that the thread
is never woken up due to this check:

        wait_event_freezable(err_detector->esd_thread_wq,
            (dsi_config->dsi_hw_context.panel_on ||
             kthread_should_stop())

When mdfld_dsi_error_detector_wakeup() is called to wake up the
thread before setting panel on to true, wait_event_freezable()
will continue to wait since it will fail the panel on check.
Moving the wakeup after setting the panel on
state prevents that from happening.

Change-Id: I8201a9a15c9088e382e0e472411522e4fc4faa8a
Signed-off-by: Vinil Cheeramvelil <vinil.cheeramvelil@intel.com>
Reviewed-on: http://android.intel.com:8080/69647
Reviewed-by: Geng, Xiujun <xiujun.geng@intel.com>
Reviewed-by: Ai, Ke <ke.ai@intel.com>
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>