atomisp2: cleanup the watchdog code
BZ: 61572

This patch removes the watchdog code and replace it with the simple
wait_for_completion_timeout() in atomisp_work.
This patch helps to reduce the complexity of file injection
patch, because in atomisp_work, watchdog must be stopped when there
is no frame buffer enqueued into CSS, and atomisp_isr also needs to
check if watchdog is alive before petting dog.

Change-Id: I307353b872a4c645122168f6fae8de2f7cbff0df
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/69659
Reviewed-by: Toivonen, Tuukka <tuukka.toivonen@intel.com>
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Reviewed-by: Laakso, Antti <antti.laakso@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>