PM: i386_ctp_defconfig: Make Interactive governor by default on CLV
BZ: 63077

The patch selects Interactive governor as the default governor for
CLV as Interactive governor proves to be the best governor on JB
currently.
This is done as the first step to the evolution of governors
on IA. The tuned IG is an intermediate step and the target is to finally move
to a UX+ govenor.

Change-Id: I99d6a63bbe7b4f8c31e506c79691ffc1147b33e9
Signed-off-by: Kumar Mahesh <mahesh.kumar.p@intel.com>
Reviewed-on: http://android.intel.com:8080/70594
Reviewed-by: Sabatier, Sebastien <sebastien.sabatier@intel.com>
Reviewed-by: Rudramuni, Vishwesh M <vishwesh.m.rudramuni@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Cuesta, FernandX <fernandx.cuesta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>