[System]: avoid warning when do ioremap for ram console
BZ: 47485

There is a warning when do ioremap for the iomem resource for Ram console.
because ram console already reserve the memory on System RAM.
But __ioremap_caller don't allow anybody to remap normal RAM that we're using, so it report a warning.

for we already reserve RAM resource for ram console below the HIGHMEM zone, so can directly map it to kernel
space to use. don't need use ioremap.

Change-Id: I54ae2e0d9d222efcc125f356f24e8d07639d39d7
Signed-off-by: Shi Yang <yang.a.shi@intel.com>
Reviewed-on: http://android.intel.com:8080/70804
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>