audio: Fix for Codec power off after the first playback during the multistream playback.
BZ: 49949

Codec is getting powered off, when the capture stream is still active.

The codec should not be powered off, when the capture stream is active.
During multi stream playback, as soon as the first playback ends, the
snd_pcm_close() will be called and since playback counter is still active,
the codec gets powered off even when the capture stream is active, due to a
wrong condition check in the soc core.

This patch does not power off the codec when the playback or capture streams
are still active.

Change-Id: I81436b9b5425c6733fed997b3b5e5a8dd355cae0
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/71122
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: P C, SreekanthX <sreekanthx.p.c@intel.com>
Tested-by: P C, SreekanthX <sreekanthx.p.c@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>