usb:gadget: Fix DMA mapping warnings
BZ: 61755

Issue 1:
The dma addr of the buffer of dev->status_req is retrieved with
virt_to_phys(), which is not compliant with DMA mapping API.

Issue 2:
When cdev->req's length is 0, langwell_ep_enqueue() maps DMA
as the length is 1, this causes length dismatching when unmapped.

The DMA mapping solution below should fix issues above.

Before calling queue_dtd():
If the length of request is 0, {dma, mapped} must be set to
{DMA_ADDR_INVALID, 0}, and no DMA mapping is needed.
otherwise, driver should map DMA if needed.

Change-Id: Ic9be5e2a75af4a97a4d42afe790b82144c1cf84e
Signed-off-by: jzhuan5 <jin.can.zhuang@intel.com>
Reviewed-on: http://android.intel.com:8080/71444
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Meng, Zhe <zhe.meng@intel.com>
Tested-by: Meng, Zhe <zhe.meng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>