atomisp: make ATOMISP_IOC_G_DIS_STAT non-blocking
BZ: 55065

Change the reading of DIS statistics non-blocking and more
uniform with the way how 3A statistics are read. Drops the
wait condition specificly added for reading the DIS statistics.

Adds checks for the streaming state and the state between
streamon and the first ISP interrupt with statistics ready
flag.

Change-Id: If1b1ed3997ad7e6e14d40579b9da8f652f710ef8
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/71883
Reviewed-by: Cohen, David A <david.a.cohen@intel.com>
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Reviewed-by: Govindapillai, Vinod <vinod.govindapillai@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>