extcon: register for cable intereset by name
BZ: 64519

There are some scnearios where a driver/framework needs to register
interest for a particular cable without specifying the extcon device
name. One such scenario is charger notifications. The platform will
have charger cabel which will be bound to any extcon device. It's
not mandatory for the charger driver to know which extcon device
it should use. This patch enables the support for registering
interest for a cable just by cable name wihtout specifying the
extcon device name

The extcon_register_interest_cable_byname() was removed after
review in mainline. The new patch modifies the existing
extcon_register_interest function to do the same.
The patch was merged into the mainline and the SHA1 id is 4f2de3bf

Signed-off-by: Jenny TC <jenny.tc@intel.com>

Change-Id: Ifc5d2031c3d8187a59f6d640000735d364713347
Reviewed-on: http://android.intel.com:8080/71634
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>