Audio-HDMI: Fix double freeing of ALSA preallocate buffer
BZ: 60964

hw_free is called two times which causes the preallocated buffer to be
freed two times, thus giving a warning message in the dmesg.

In alsa-lib, hw_free is called as an ioctl call during snd_pcm_close and
again it is called inside the alsa kernel release call. This cause the
driver's hw_free to be called two times, which attempts to free the
buffer two times.

Added a check in driver hw_free to track the double freeing up of the
preallocated buffer.

Also removed some unused variables.

Change-Id: I36553e61b0ab53e04da47ab3e91785acab9b1e64
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Reviewed-on: http://android.intel.com:8080/72295
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: P C, SreekanthX <sreekanthx.p.c@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>