Display: Fix green bar on screen when stop system_server
BZ: 65329

This bug is due to the stop of system_server does not call destruction
function of SurfaceFlinger and HWComposer, and thus overlay could not
be disabled in the process. When stop system_server, the swapchain will
be destroyed, so fix the issue by flipping screen with init_screen buffer
to hide the overlay.

Change-Id: Idcebacbf243dc2b7ed6e723d8a88f5abb0bffd0f
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/72134
Reviewed-by: Mao, Marc <marc.mao@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>