atomisp2: implements new locking/serialization for atomisp2
BZ: 60510

Adds new mutex that will be used for protecting atomisp2 driver.
When user space calls ioctl this mutex is taken. Same lock is also taken
when driver access its own structures or call css API.

This might potentially harm performance but because there really is
only one HW this is right way to do it with current API.

Change-Id: I98b380ca17287f1b4db54f2d255fa7a126d9eb78
Signed-off-by: Jukka Kaartinen <jukka.o.kaartinen@intel.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/72020
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Tested-by: lab_cactus <lab_cactus@intel.com>