atomisp2: reshuffle stuff in probe to remove redundant flag
BZ: 63963

The sw_contex.probe flag is redundant. It was used as a cheap lock in a hackish
way. The device registration should be done at the end of the probe() function.
That's why this patch moves it there.

Additionally it removes unnecessary check "if (!isp)" in the atomisp_open().
Accordingly to Sakari it could be NULL if video_set_drvdata() wasn't called in
atomisp_video_init() which is not going to happen. video_set_drvdata() is
called before the device nodes are created.

Change-Id: I94c89a3322d10a676a5ac71ad8b6b7e487c88c51
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Jukka Kaartinen <jukka.o.kaartinen@intel.com>
Reviewed-on: http://android.intel.com:8080/72885
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>