ov8830: set the default binning option as "average" for all modes
BZ: 67078

The user space library expects the binning ooption used as average
when binning is used. So set the default valur for binning option
as "average" in sensor's basic settings list and remove the
"sum" binning option settings from the modes which uses binning.
This will solve the issue of brightness level being different in
preview and still capture modes.

Change-Id: Ic136d1a9ef7bbaf439d101bf3640da92a41c2e70
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/74048
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>