dw-i2c: Correct the timing to issue STOP on TNG
BZ: 67637

On TNG, I2C controller have to issue a STOP after each transaction.
Currently, the STOP is issued at the last command of every I2C message,
which is incorrect and may cause interrupt disorder.
The STOP should be issued at the time of last command of the last I2C message.
This patch fixed this bug.

Change-Id: I3f5034ac238cb804ae8d3504c0c5e799cd3f0ff2
Signed-off-by: Li, Ning <ning.li@intel.com>
Reviewed-on: http://android.intel.com:8080/73549
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>