dw-i2c: reduce I2C driver RX_FULL interrupts
BZ: 67637

I2C controller generates too many RX_FULL interrupts when using
touch, which is rootcaused as the threshold set in I2C_RX_TL register
is always zero.

The solution is to dynamically setting the threshold according to
the I2C message length and checking the RX FIFO whenever a STOP
interrupt comes, to ensure the data is fully read.

It is tested on MFLD_PR2 and CTP_PR1 that the number of RX_FULL
interrupt for each touch action could be greatly reduced.

Change-Id: I5d856230590235a3bc8984df5db54994d4837713
Signed-off-by: Li, Ning <ning.li@intel.com>
Reviewed-on: http://android.intel.com:8080/73550
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>