[BT][KERNEL] check NULL pointer deference at __sco_sock_close function
BZ: 65292

A kernel panic due to a NULL pointer deference has been signaled for
this function. It's a very rare event that probably happens because of
a concurrence issue is not solved for BT sco socket for old kernels.

This patch simply checks if the pointer is not null before using it.
It prevents the NULL pointer deference, but does not solve the
concurrence issue (that has been probably solved in following version
of the kernel, where everything in HCI run in process context, and
mutex are used instead of spin_lock).

notes:
1) commit f21ca5fff6e548833fa5ee8867239a8378623150 (Von Dentz-Padovan)
   fixed shutdown on SCO sockets
2) commit ce8453776d68982cfe93bcb28191af8ccad01f45 (Torvalds)
   reverted the commit f21ca5fff6e548833fa5ee8867239a8378623150 with
   the reason "... can cause a NULL dereference ..."
3) commit 0da67bed835fdde68ca0e924d2a2d6ac82c70833 (Von Dentz-Padovan)
   applied again the patch f21ca5fff6e548833fa5ee8867239a8378623150
The patch proposed by Von Dentz-Padovan makes sense even if, in
Torvalds' mind, it can cause a NULL pointer deference.
So we keep the latest version, adding the check for the NULL pointer.

Change-Id: Ia59c72188973768107fc03992b1a28cfcec3755f
Signed-off-by: Mirko Biselli <mirkox.biselli@intel.com>
Reviewed-on: http://android.intel.com:8080/74182
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Bondier, CedricX <cedricx.bondier@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Reviewed-by: Bransiec, ChristopheX <christophex.bransiec@intel.com>
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>