[PORT FROM R3]mmc: sd: retry one time if encounter failure to switch 1.8 voltage
BZ: 67120

If a SD card support to switch to 1.8v, SD driver will send CMD to do so.
But if the SD card actualy failed to do the switch, SD driver will send CMD
again and again until the SD success to do so.

That is not safe for the SD card which has issue to switch voltage. So add a
check before sending switch command again and again. Just let the SD driver
to retry one time. And if the SD card failed to switch, continue to use it
as the original voltage.

Change-Id: I9db0ebd5f0459007387f1fe487694d64b14d8067
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/74543
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>