trace/runtime_pm: Fix the crash due to NULL dev_name
BZ: 68482

Commit 59b8a6bf8b350(trace/runtime_pm: add runtime_pm trace event)
introduced the runtime_status trace event, also intrduced this issue.

This crash call stack is as below:
    mmc_rescan
    -> mmc_rescan_try_freq
    -> mmc_attach_sdio
    -> pm_runtime_set_active
    -> __pm_runtime_set_status
    -> __update_runtime_status
    -> trace_runtime_pm_status
      This function is corresponding to the below code in trace/events/power.h:
    TRACE_EVENT(runtime_pm_status,

        TP_PROTO(struct device *dev, int status),

        TP_ARGS(dev, status),

        TP_STRUCT__entry(
                __string(devname, dev_name(dev))

In sdio.c, the dev_name is set at mmc_add_card(), so when pm_runtime_set_active()
is called, at this time the dev_name is still NULL, it causes this issue.
Of course, this issue is not always reproduced, it is necessary the trace event
must be enabled.

So here fix it that changing TRACE_EVENT to TRACE_EVENT_CONDITION, to avoid this
panic case that dev_name is NULL and trace event runtime_status is enabled, but
pm_runtime_set_active is called.

Change-Id: I9803b834304cbdfb8891fd9545367b40a1846bb3
Signed-off-by: liu chuansheng <chuansheng.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/74329
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>