Add timer to monitor that suspend worker thread is not stucked.
BZ: 62530

Suspend worker thread is supplied with 3 types of works : early suspend,
late resume and suspend. This patch adds one timer by work: we start the
timer at the time the work is queued in suspend workqueue and it's stopped
at the time the worker begins to execute work's function. If, for some
reason the suspend worker is stucked, at least one timer will expire and
the kernel will BUG().

Change-Id: I14361c9f7601e05c8330ce483f986729dd5179ad
Signed-off-by: Catalin Popescu <catalinx.popescu@intel.com>
Reviewed-on: http://android.intel.com:8080/74719
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Sil, Dyut K <dyut.k.sil@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>