atomisp2: initialise IUNIT registers values on probe and save those
BZ: 65408

The current functions in atomisp2 to save and restore the IUNIT
register settings is moved to atomisp_v4l2.c and the save function
will be called only when the device boot up. And restore function
will be called on every camera start-up. This is part of the
the patch which ensures the correct register setting for the
IUNIT registers which has an impact on the ISP performance.

Change-Id: I24d45166a2a5c7ee4b49bc53ce4ce7bf08ff1a04
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Acked-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/76122
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Tested-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>