Improve the load compuation for CPUFreq governor
BZ: 69548

This is patch #1 of intel governor patch series

In hyper-threading enabled system, the cpu load obtained in CPUFreq
governor can only reflect the load of logical core, which does
not reflect the real load of the physical core, so that the
incorrect cpu load leads to incorrect frequency adjustment.

This patch introduces a method to compute the physical core
load for hyper-threading enabled system. Our measurement shows
that the patch can help to improve the performance and power of
typical android applications.

Change-Id: I2b0ea72f6260106643161a061e79cb4d644016cb
Signed-off-by: cwang99 <chuan.a.wang@intel.com>
Signed-off-by: ke <ke.chen@intel.com>
Signed-off-by: li <li.w.wang@intel.com>
Signed-off-by: xiao-feng <xiao-feng.li@intel.com>
Signed-off-by: Sundar Iyer <sundar.iyer@intel.com>
Reviewed-on: http://android.intel.com:8080/76082
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>