Fix "No video indentification code available"
BZ: 54618

When searchng VIC for a given video timing mode, instead of comparing clock value for 60Hz or 59.94HZ to the given clock,
clock range between 59.94Hz and 60Hz is used to check whether the given clock value is allowable, this provides more leeway
when clock is round-up to different precision.

Two patches are needed to fix BZ 54618. This is patch #2.

Change-Id: Id002c9a30d8e4aed6bca7d2fce3946a2746fe0cb
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/76400
Reviewed-by: Jin, Fu <fu.jin@intel.com>
Reviewed-by: Liu, Wei W <wei.w.liu@intel.com>
Reviewed-by: Jia, Lin A <lin.a.jia@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>