GFX: making sure the interrupt of HWR is handled as early as possible
BZ: 71997

This patch is to make the HWR interrupt handling as early as possible,
it can reduce the dirty data to be produced.

Thru our tests, we found we can not do HWR recovery at any time, it must
hit the below points:
1/ Ukernel is in idle;
2/ Ukernel has gone into dealock(lock,lock) state;
3/ Ukernel is finished the powering off action;

Even in the above cases, Ukernel will be powered on again and do the
recovery(syncing object...) work. So we do not want to many dirty
in the shared ram to avoid HWR failure.

Change-Id: Iafb2367755460649b5e8e72c81f4084fa337db59
Signed-off-by: liu chuansheng <chuansheng.liu@intel.com>
Signed-off-by: Tong, Bo <box.tong@intel.com>
Reviewed-on: http://android.intel.com:8080/78231
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>