gfx, svrkm: replacing the msleep() to ulseep_range when timeout < 20ms
BZ: 70980

In svrkm code, there are many places using OSSleepms(), but the timeout
values are < 20ms.

According to linux kernel design, msleep() is not fittable for timeout
value < 20ms, it will cause 20ms delaying almost.

The impact is it should cause the performance decreasing and power effect,
in function PVRSRVPowerLock(), OSSleepms(10) is called which may cause
20ms delaying indeed when handling CCB command request.

Change-Id: Ibbc4c3c0d796eb2cb9b044eb0444f00b1dca1c98
Signed-off-by: liu chuansheng <chuansheng.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/78229
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>