Drivers: Staging: Sep: send shared memory address to device
BZ: 60970

The sep device needs to know when we are in operating system
operation and that the SCU and x86 firmwares are done with their
boot-ups.

The sep device also needs to know the physical address of the
shared memory region.

This patch will accept the non-valid-opcode error from the device;
this is in recognition that not all firmware will have this
implimented.

Change-Id: Ia8c48642140782a53ada5dee4f26cf80321cfc18
Signed-off-by: Mark Allyn <mark.a.allyn@intel.com>
Reviewed-on: http://android.intel.com:8080/72064
Reviewed-by: Matthews, Michael L <michael.l.matthews@intel.com>
Reviewed-by: Shuster, Norvin <norvin.shuster@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Mangalampalli, Jayant <jayant.mangalampalli@intel.com>
Tested-by: Mangalampalli, Jayant <jayant.mangalampalli@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>