vibra: enable mrfld vibra driver
BZ: 46529

This patch enables the LRA based vibra for mrfld. The SoC PWM controller needs
to give PWM output >50% duty cycle to TI DRV 2605. The driver needs to be
calibrated at boot time for the LRA actuator. And once calibrated we need to
configure it for PWM mode.
After this enable gpio triggers driver to go from standby to active mode

The TNG GPIO is connected to PWM block in ALT_FUNC_1 unlike in CTP where it is
ALT_FUNC_2. Also the usage of driver and defaults for PWM programming are
different, so abstract these values in driver info to keep driver clean

While at it, noticed the GPIO is high on TNG after disable. We disable GPIO but
we are in ALT_FUNC mode, so put mode back to 0 and disable GPIO.
Also programming from user space doesnt work if we dont do suspend, so program
register everytime on enable.

Also move the driver to late_initcall as we depend on i2c enumration to complete
Further fix needed for moving CTP GPIOs to get_by_name and populate both thru
SFI.

Change-Id: Ib23ae94aa197cc47af4cb738c9a11060abf37f31
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Reviewed-on: http://android.intel.com:8080/79688
Reviewed-by: Jayaraman, Venkat <venkat.jayaraman@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Kalliguddi, Hema <hema.kalliguddi@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>