EM/charger: Battery Status not showing FULL issue
BZ: 71142

This patch fixes the battery not showing FULL issue. Battery takes
too long to reach FULL capacity. The reason for this is FG expects
the battery to take as low as 50mA charging current, however because
of the charger's limitation (can terminate the charging no lower than 128mA)
the charging gets terminated when the charging current drops to ~100mA
and no current flowing into the battery.

Change-Id: I3db47f0de9d2c8ecaf0ebce13bf3f172c3fb2d42
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/77832
Reviewed-by: Kannappan, R <r.kannappan@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>