audio: enable the runtime pm support for sst driver
BZ: 35783

This implements mrfld fw save and restore for audio fw. In mrfld the audio rams
(ICCM/DCCM) snapshot will be saved by driver on suspend. This snapshot will be
loaded on resume. The FW will flush the cache on receiving D3 message from
driver. FW will not allow suspend unless FW is idle, so error would mean FW is
not idle.

Temporary workaround is provided to not use the FW ram snapshot on resume but
do the complete FW download.

Change-Id: Ief1d47bbcd5458ba9f9fa9454d9f22388f752613
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Reviewed-on: http://android.intel.com:8080/78947
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>