hsi: When a modem reset is detected, send an hangup signal to clients
BZ: 63201

Because the clients don't close their connections when a modem
reset is detected, the AP continues to send data as if the
 modem is ready to receive. The modem reacts by generating a coredump.

The correction consists in requesting a client hangup when the reset_out
goes down. This makes the HSI driver clients flush their channels and
close their connections.

Change-Id: Ic2cc4073aa3462db152440461101dfc31ba83d0c
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/78733
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>