[PORT FROM K3.0]mmc: core: do not set the limitation for max_discard_sectors
BZ: 66900

If set the max_discard_sectors, erasing a big area will cause
a huge time to be finished. So remove it.

Instead, using driver to poll device status instead of let
host controller itself to do so.

With this change, the erase/trim IO time will be larger than before

Change-Id: Ie35ecc6175ae6d300efc87bfad49e105dacf432b
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/80171
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Yang, Fei <fei.yang@intel.com>