n_gsm: Rework of close procedure
BZ: 74701

This patch is to allow the mux to close properly all DLCI as gsmld_close can't :

When gsmld_close is called, tty can't be used anymore !
Here is the 3 possible cases :
      - Close of the tty on which the mux is associated
        tty->ops->close is called before tty_disc_release
      - Hangup of the tty on which the mux is assocated
        A modem/link problem occurs and modem is not reachable
      - Replacing LD by another one
        tty->receive_room is set to 0, no modem answer can be
	received

So, no need to send anything to modem, ether it will not
reach the modem, the modem can't accept it, or
we will not receive modem answer and we will timeout (7.5s).

To close modem mux, a new ioctl (GSMIOC_DEMUX) have been
added. This need to be called before the close ttyIFX0.
This ioctl close all DLCI sending DISC and waiting for UA.
This will also avoir 30s timeout from tty waiting incomming
command (DISC) to be sent to HSI link.

Now, gsmld_close only removes structures and doen't try to communicate
with the modem.

Change-Id: Ib5a4be9710c59517f1ece950ee43eda67158f19d
Signed-off-by: Vincent Pillet <vincentx.pillet@intel.com>
Reviewed-on: http://android.intel.com:8080/81394
Reviewed-by: Bourree, Gwenn <gwenn.bourree@intel.com>
Reviewed-by: Haezebrouck, Sebastien <sebastien.haezebrouck@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>