Time: Use S3 Non-stop TSC for S3 time counting instead of RTC
BZ: 53816

MFLD and CLVP have a TSC which is already running even if the system
go to the S0i3/S3 state.

Current kernel time counting will use RTC to compensate xtime after
resuming from S3, but there is some problem with MFLD SCU FW (no for
CLVP) which has a "+1" issue, and can trigger the OS time drift after
long run.

So we skip the problemtic RTC part, and just use the non-stop TSC
for time counting for S3. This code will affect MFLD and CLVP platforms
only.

We will try to push this patch to upstream soon.

Change-Id: Ica999d27ef058459b2567f6dbfa000154db5cd98
Signed-off-by: Feng Tang <feng.tang@intel.com>
Reviewed-on: http://android.intel.com:8080/81512
Reviewed-by: Du, Alek <alek.du@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>