Audio-HDMI: Fix KW issue for array index out of bound for channel_allocation
BZ: 76430

If no match is found for speaker mask derived from the ELD speaker
allocation block with speaker mask of channel_allocations, then an
invalid index of channel_allocation array may be attempted to be
assigned.

Moved the chmap and channel assignment inside the speaker mask condition
checking to avoid the array index out of bound.

Change-Id: I0ca6699766b3e6f2e8888224124ab4c678329654
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Reviewed-on: http://android.intel.com:8080/83391
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Tested-by: P C, SreekanthX <sreekanthx.p.c@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>