max17042: add fake_batt_full option
BZ: 48314

Some platforms like pnp boards report a low capacity
even when they are connected full time to
a charger and they dont have a battery.

Android will force a reboot when it detects a low capacity
on the battery.

To make it easier to work with these platforms and
to prevent unnecessary constant reboot add a boot arg option
to fake a full battery.

Change-Id: I78f056d85d965756141a84acc3ee2e8f84493894
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Signed-off-by: Ramakrishna Pallala <ramakrishna.pallala@intel.com>
Reviewed-on: http://android.intel.com:8080/84652
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: Mitra, DebalinaX <debalinax.mitra@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>