bcm433x: enable bluetooth Low Power Mode
BZ: 82570

Enable bluetooth Low Power Mode (LPM) by
using the 'wake_peer' callback to enable the
blueooth chip via a GPIO and enable the UART using
runtime_pm.
Also avoid glitch with rfkill and set
bluetooth off on init.
The GPIO's are checked with 'gpio_is_valid' to avoid
crash on boot if the GPIO's are not present
in the SFI table.
To make the driver more portable, get
the UART port number from the platform data.

Change-Id: I0820f1f3ceca068d3bc440ac1b434c29d02d0799
Signed-off-by: Frode Isaksen <fisaksex@intel.com>
Reviewed-on: http://android.intel.com:8080/84996
Reviewed-by: Isaksen, FrodeX <frodex.isaksen@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>