audio: mrfld_machine: add jack detection support
BZ: 77856

Add jack detection support to Merrifield boards. The codec interrupt number is
received from platform data and a jack is registered with ASoC. A callback
handler checks the jack status whenever an interrupt is triggered. The interrupt
registers in the LM49453 codecs are 'read-clear' and the codec interrupt line
will remain high till all the unmasked interrupt registers are cleared.

The jack detection logic uses the jack type reported by the codec to report the
jack to the upper layers.

Also fixup the machine driver to have card data set before registration so that
it is availble at dailink init.

Change-Id: I4774c6f0ddba53d0c561195371b2055b19cf3ecb
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/85680
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>