Enable hung task
BZ: 79361

IPANIC at suspend_timeout_fn+0x17/0x20 is seen after patch 74719 merged,
this patch add a timer which intended to detect suspend worker thread stuck.
However, this patch has some design issue which will trigger the IPANIC
unexpectly. So it is reverted in bug 77706.

Since kernel already has a mechanism to track task hung, it creates a kthread
and periodically check whether a thread is stuck in UNINTERRUPTIBLE state too
long. So we can use this mechanism to detect all the deadlock or wait forever
issue, including the stuck of suspend worker thread.

This patch is to enable the task hung feature. The check period is set to 60
seconds and it will panic on finding the hung task.

Signed-off-by: Zhang Di <di.zhang@intel.com>

Change-Id: I8dc37fc06aea253116772a6cdc894830690264d2
Reviewed-on: http://android.intel.com:8080/85738
Reviewed-by: Liu, Chuansheng <chuansheng.liu@intel.com>
Reviewed-by: Popescu, Catalin <catalin.popescu@intel.com>
Reviewed-by: Zhang, Di <di.zhang@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>