[PORT FROM R4.1] mt9d113: adjust default gain value to avoid overexposure issue.
BZ: 78962

Issue:
The preview is overexposure when switch to preview/panorama from
video. This is because that the default gain value set in preview
mode is so small, so exposure will be over.

Solution:
Adjust the default gain value to avoid this overexposure issue.

In additional, a defect of reading reg in mt9d113_get_iso function
also is corrected.

Change-Id: I888505c1896ce43baa3a58196ce8abce9ad086b5
Orig-Change-Id: Ia52f5c8f906e73c5f5fb7b61c29e421169da5200
Signed-off-by: zjiang2x <ziyux.jiang@intel.com>
Signed-off-by: Tao, Jing <jing.tao@intel.com>
Reviewed-on: http://android.intel.com:8080/86230
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>