wake up the kintel_up/kinteractiveup thread after create
BZ: 79361

The thread kintel_up isn't waken up after kthread_create, so it is always in
UNINTERRUPTIBLE status if no one use it.
khungtaskd can detect the tasks who stuck in this state for a certain time,
so thread kintel_up is caught by task hung.

The solution is wake up the thread after kthread_create so that it will in
correct status and khungtaskd won't complain about it.

The thread kinteractiveup has the same problem, so fix it together.

Change-Id: Id4e7d2a1429746f4327bdfdc4ff28b5b4834a600
Signed-off-by: Zhang Di <di.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/85992
Reviewed-by: Liu, Chuansheng <chuansheng.liu@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>