[PORT FROM K3.0] atomisp: remove the dependency of camera module presence from probe
BZ: 65868

atomisp_pci_probe() must successfully complete even if there is no
physical camera module attached to board. This is required to power
off the ISP block in case of COS mode and also in case the dev
board contains no camera modules.

This is for the old atomisp driver used in medfield based devices

Change-Id: I5878c88d5f2a5bd92e96949dc719dbd076000ccf
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Reviewed-on: http://android.intel.com:8080/86076
Reviewed-by: Popescu, Catalin <catalin.popescu@intel.com>
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-on: http://android.intel.com:8080/88024
Reviewed-by: Cohen, David A <david.a.cohen@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>