ov8830: updating ov8835 modes as per Omnivision recommendation
BZ: 80103

Updating the ov8835 sensor modes as per the omnivision
recommended settings.

Skipping is not used in any of these current modes anymore.
Binning factor information in the reolution table is updated
as per the correct setting used in the register list.
Modified 1080P DVS resolution setting
Moved the R3810-R3814 to common settins as they are not changed
in the mode specific settings any more.

Change-Id: If79958b71b38eaf1ef47495c2648ae720aeb6656
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/87078
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>