[PORT FROM K3.0] ov8830: separate frame timing settings from the register list
BZ:80103

No need to include the frame timing register as part of the
register lists for each mode. Instead this can be loaded from
the resolution  struct when a new mode is selected. As the
exposure affects the frame timings, exposre setting is combined
with the frame timing settings and they are written to the
register as a one block.

Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>

Removed redundant "goto out".

Change-Id: Idf6bd3a7a89d6949d606bf0c8e9479fbca0c838e
Orig-Change-Id: I673ce87b652f19d235a230271a6cbeefe9e955f8
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/90052
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>