audio: SFI entry name change for the CTP machine driver
BZ: 81676

SFI entry for the CTP machine driver is changed from clvcs_audio to ctp_audio in
the IFWI. So, the occurance of this string need to be changed in the board file
and in the ctp_common machine driver for the successfull driver enumeration.

This patch changes the machine driver name from clvcs_audio to ctp_audio.

Change-Id: I23ddfbb74bb157823aeca623b0eb1cf170025aec
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/87678
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: Continente, Christophe <christophe.continente@intel.com>
Reviewed-by: Chotard, Celine <celine.chotard@intel.com>
Tested-by: Chotard, Celine <celine.chotard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>