[PORT FROM K3.0] ov8830: Modified the g_framinterval call.
BZ: 78914

g_frmainterval call must return the currently selected sensor
settings' frame interval. This will be the maximum FPS this
setting can produce. But the "actual" FPS is based on the
exposure setting applied. But the user side is providing the
exposure settings, it should know what exposure setting is being
applied and calculate the accurate FPS at any given point of time.

Change-Id: I376b24319c65d0edf1dcf437a75700c1eec9a064
Orig-Change-Id: I4e4a1cabb3d2728a55627a6a82b5fed392c03136
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Acked-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/90603
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>