MCD: Power off 7x6x modem at boot
BZ: 82377

As confirmed by SE team, powering off the modem can only be done with 7x60 and
6360 modems.
For now, this is handle by KCONFIG variable.

Early boot is keeped on MFLD platforms  as it is mandatory to not keep it off.

All platform specifities will be handled with a platform device in an other
patch.

Change-Id: If26743733f869e5f0c212b13e10436461261019f
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>
Reviewed-on: http://android.intel.com:8080/88725
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Kouassu, Alain <alain.kouassu@intel.com>
Reviewed-by: Robert, Denis <denis.robert@intel.com>
Tested-by: Robert, Denis <denis.robert@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>