[BT] [PATCH 03/16] Enable HID profile for Bluedroid (UHID)
BZ: 83506

From: David Herrmann <dh.herrmann@googlemail.com>
Date: Sun, 10 Jun 2012 15:16:15 +0200
Subject: [PATCH 03/16] HID: uhid: allow poll()'ing on uhid devices

As long as the internal buffer is not empty, we return POLLIN to
user-space.

uhid->head and uhid->tail are no atomics so the comparison may return
inexact results. However, this doesn't matter here as user-space would
need to poll() in two threads simultaneously to trigger this. And in this
case it doesn't matter if a cached result is returned or the exact new
result as user-space does not know which thread returns first from poll()
and the following read(). So it is safe to compare the values without
locking.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>

Change-Id: Iaa99ec61ed5e65b1cb2956e878559921fa8cbeb4
Signed-off-by: Cedric Bondier <cedricx.bondier@intel.com>
Reviewed-on: http://android.intel.com:8080/89162
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>