REVERTME: Temporaily disable IDLE and PMU driver for merr_vv
BZ: 83781

Currently earlyprintk is enabled by default for merr_vv build,
this is causing problem for PMU driver that the PMU probe function
would timeout as max3111 is on for earlyprintk. This patch temporarily
disables CONFIG_INTEL_IDLE and CONFIG_INTEL_ATOM_MRFLD_POWER so that
droidboot could boot up completely on merr_vv to unblock teams on porting
patches for kernel 3.4.
IDLE and PMU driver needs to be re-enabled once real solution of this
issue is ready.

Change-Id: Iaf29a97e196af8803c960939464677b15135682a
Signed-off-by: Fei Yang <fei.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/89295
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>