Display: correct mipi overlay flip legal check
BZ: 83833

This patch is to fix the green display on MIPI when plug in HDMI while
playing video during a incoming call. It was supposed that when HDMI
is connected, overlay should always be switched to pipe b for extend
video mode. But it is not true under the incoming call case. So need
to correct the flip legal check.

Change-Id: I3301db355df014aee423b0a84e1102bfb2f94b4d
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/89750
Reviewed-by: Liu, Hong <hong.liu@intel.com>
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>