Audio: Dynamic block support added
BZ: 71391

Currently static blocks are used to block for response from firmware, fw
download etc. This results in difficulty tracking the block assignment
and response. Also resulted in race condition in one of the scenario in
a multithreaded environment

Changes are added with which the blocks are created dynamically and
freed. This makes the code much cleaner and easier to track the
response

Change-Id: I9dd67e0d7637111be247fdbaed7128275f350bf1
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Reviewed-on: http://android.intel.com:8080/89599
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: P C, SreekanthX <sreekanthx.p.c@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>