ov8830: Skip first frame from sensor for preview modes in OV8835
BZ: 79168

The raw image shows the first from the sensor has green levels high
And this is getting stabilised from the second frame onwards. So
until we find a proper solution for this issue, it has been decided
to skip first frame from the sensor. The frame is skipped from the
Camera HAL based on the information provided by the sensor driver.

Change-Id: I57de03ecbae28333481725ee4157a720b7ee2f4c
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/90785
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: Kontola, Kalervo <kalervo.kontola@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>