st_kim: enable runtime PM on the tty device used by st_kim driver.
BZ: 84667

In order to load the firmware, the tty_dev must be in active state
while loading the firmware. To do that, st_kim must wait for the line
discipline to be installed, then it can identify the tty device
involved.

Blocking delays to reset the chip have been replaced by their
corresponding sleepy funtions.

And finally, the timeout used by the st_kim to wait for the user space
process to install the line discipline has been adjusted to 2 secs.

Change-Id: Ice287067c7105a6aba53021e680f499f81b426ba
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/90962
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Bondier, CedricX <cedricx.bondier@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>