[PORT FROM K3.0] audio: Dynamic Configuration of platform specific configuration in LPE
BZ: 85347

This patch provides facility to dynamically configure the LPE firmware
for different platforms just by enabling the platform specific
configuration in the board file.

The platform specific configuration is initialized in the board file.
This information is passed to the SST driver, during the SST driver
registeration.SST driver writes the LPE specific configurations to DRAM
region from the board file. This enables the LPE firmware to dynamically
configure the board and platform specific configurations by reading the
configuration region from the DRAM region.

Change-Id: Ie00f0053223d781a3984510e52a765e5750cb0fc
Orig-Change-Id: Icd6428cf3c6feddc24d18b001e860b01fa18aa13
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/91639
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: P C, SreekanthX <sreekanthx.p.c@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>