audio: Dynamic Configuration of platform specific configuration in LPE
BZ: 85347

This patch provides facility to dynamically configure the LPE firmware
for different platforms just by enabling the platform specific
configuration in the board file.

The platform specific configuration is initialized in the board file.
This information is passed to the SST driver, during the SST driver
registeration.SST driver writes the LPE specific configurations to DRAM
region from the board file. This enables the LPE firmware to dynamically
configure the board and platform specific configurations by reading the
configuration region from the DRAM region.

Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Change-Id: Icd6428cf3c6feddc24d18b001e860b01fa18aa13
Reviewed-on: http://android.intel.com:8080/90266
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>