[PORT FROM K3.0] mid_pmu: add sysfs to sync os seen d0ix with SCU
BZ: 86874

Since by default all LSS are ignored in PMU driver,
When LSS are re-enabled, if an LSS was in D0i3 from OS point
of view, and does not issue an D0ix command for long time,
the OS view and SCU view of SSS will be in difference.

Hence providing a sysfs interface to sync the differing LSS
if they are allowed to be sent to SCU.

This will send D0ix command if the OS view and SCU view of
allowed LSS are different.

Change-Id: Ie31f1543e51928355619371601d4991b4d3e83e1
Orig-Change-Id: I825408d70b78b8041a442eab34eb25e51b72b0c6
Signed-off-by: Illyas Mansoor <illyas.mansoor@intel.com>
Reviewed-on: http://android.intel.com:8080/92375
Reviewed-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>