[Telephony/HSI_eDLP]: Wake up the read waitqueue to unblock the poll function before closing
BZ: 86001

The close of the mdmTrace is blocking by the client of this
stream driver that waits on POLL function. The only way to
finish the close and to allow another client to open a new
instance of the driver is to unblock the poll_wait. To do
this, the read waitqueue must be woken up and a flag must
be used to go out of the poll function with a specific event.
The chosen event here is POLLHUP.

Change-Id: I97bba9750d7b948cdb61b886c921fa4ba4fb0656
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/92130
Reviewed-by: Tenoutit, FaouazX <faouazx.tenoutit@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>