pti : support of MRFLD
BZ: 87375

PTI support is not present for Intel tangier SoC, due to new PCI enumerations.

The new PCI ressources, created for Intel tangier SoC, are provided through
Simple Firmware Interface (SFI). The PCI ressource index to use is chip-dependent.

A function is added to known which PCI ressource to use for register accesses.

Change-Id: I71fba87826ca9109ba812ef97d81d2d3310912b8
Signed-off-by: Laurent JOSSE <laurent.josse@intel.com>
Reviewed-on: http://android.intel.com:8080/92403
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Hogander, Jouni <jouni.hogander@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>