[PORT FROM K3.0] EM/Charger: remove polling method and interrupt threshold correction
BZ: 77555

The patch resolves removing of polling mechanism while charger connected.
While Charging the % increase is not updated by the max17047 chip due to
rounding of the capacity as 93.5%->93+1, so the soc_threshold is incremented
to next higher value which is never crossed. Removed the rounding off of
capacity as 93.5%->93 which resolves the issue.

Change-Id: Iacdd351d8199a46908269e438db3cf125ed2d436
Signed-off-by: Naveen Rapaka <naveen.rapaka@intel.com>
Reviewed-on: http://android.intel.com:8080/92548
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Kannappan, R <r.kannappan@intel.com>
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: Kundella, SreedharX <sreedharx.kundella@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>