[Telephony/Hsi_eDLP]: Solve the kmemleak issues in the dlp_ctrl_cmd_send function
BZ: 87494

In the dlp_ctrl_cmd_send function there are several ways
to go out. When an error happens before the sending of
the eDLP command, only the previous memory allocations
must bee freed, but after sending the command to the
controller, all the PDUs, command area and frame memory
area must be freed. The branch to be used in the different
cases have been updated according to the memory allocations
done. Some comments have been added to indicate where the
different frees have been done if not in this function.

Change-Id: Ie22d5ee88d21e20cad6cff4284d518da8e502f41
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/92586
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tenoutit, FaouazX <faouazx.tenoutit@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>