Audio: Debugfs: Dump Iram and Dram when recovery is triggered
BZ: 84611

When the sst recovery is triggered, the iram and dram contents are saved in
local buffers and an uevent is issued to Userspace.

The userspace can mmap the iram and dram buffers through the following interface:
    1) d/sst/iram_dump
    2) d/sst/dram_dump

The uevent contains events:
    IRAM_DUMP_SIZE=<size of iram>
    DRAM_DUMP_SIZE=<size of dram>

These can be used by the userspace to get the specific sizes of the buffers to be read.

Change-Id: I26e91b45677238c05bea4eac923afd3075d742cf
Signed-off-by: Samreen Nilofer <samreen.nilofer@intel.com>
Reviewed-on: http://android.intel.com:8080/89830
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Beaumont, PhilippeX <philippex.beaumont@intel.com>
Tested-by: Beaumont, PhilippeX <philippex.beaumont@intel.com>