gpio-key: free irq before enter S3
BZ: 85506

As there is no mask/unmask support on CTP for gpio irq, it need use
free_irq to disable irq when the gpio is not a wakeup source.
We also need some gpio can wakeup system in lpmp3 mode, so export
enable/disable wakeup sys interface to user space.

Change-Id: Icb9e527ccaed8b419f099703eeb812f0e7fcf7f7
Signed-off-by: Chen, Jason <jason.cj.chen@intel.com>
Reviewed-on: http://android.intel.com:8080/93181
Reviewed-by: Li, Ning <ning.li@intel.com>
Reviewed-by: Yang, Bin <bin.yang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Reviewed-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>