[Merrifield]: No video content displayed during playback via gallery.
BZ: 88328

Video driver failed to creat buffer base on user pointer of graphics buffer,
the reason is that pages of these buffers are PAGE_SPECIAL which prevent video driver
to get user pages. so replace remap_pfn_range with vm_insert page, in which
PAGE_SPECIAL will not be set.

Change-Id: I693bd15d5ff7a865326b18c8680a27c0c2c713b7
Signed-off-by: mahongpe <mahongpe@intel.com>
Reviewed-on: http://android.intel.com:8080/93513
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ma, Hongpeng <hongpeng.ma@intel.com>
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Zeng, Li <li.zeng@intel.com>
Reviewed-by: Dai, Yu <yu.dai@intel.com>
Reviewed-by: Guo, Nana N <nana.n.guo@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>