[Telephony/MCD] Create XMM6360/7160 power on sequence
BZ: 86458

As an inheritance from eDLP, the power on sequence was a cold reset sequence,
that handles any previous and unexpected state.
The aim here is to be compliant with last versions of SAS, that prefer
a specific sequence, considering that modem SHOULD be off once power on
sequence is being exectuted.

Change-Id: Icc68d800bc9e24bf23f05409c3a63ace16e0e896
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>
Reviewed-on: http://android.intel.com:8080/94061
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Tenoutit, FaouazX <faouazx.tenoutit@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>