PM: Implement sys_sync() in other works during suspend
BZ: 81278

Currently, in enter_state() during suspend, it will call sys_sync(),
but in some case, the sys_sync() need more time to be finished.

For this case, if user press the power button to wakeup the system,
it will not succeed, because the suspend_lock is held during sys_sync,
therefore there is no chance for late_resume_work to be enqueued.

So here we put sys_sync() in another new work to be finished, meanwhile
during suspend, will check the wakeup pending every 200ms, once new
wakeup pending, suspend will be break and wakeup process can continue.

Change-Id: Ifa5e5122a0d1040d1e92917087929ea65097ebab
Signed-off-by: Li, Fei <fei.li@intel.com>
Signed-off-by: liu chuansheng <chuansheng.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/88041
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Dumur, CharlyX <charlyx.dumur@intel.com>
Reviewed-by: Popescu, Catalin <catalin.popescu@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>