Audio: Fix for headphone detected as headset
BZ: 87636

Reading the MIC2 short detect immediately after enabling mic bias causes
the headphone detected as headset.

State transitions for MIC2_SDET pin is debounced for 20ms in both rising
and falling edge direction, this may cause incorrect reading of this pin.

Keeping the mic_bias always on if it is detected as the headset/headphone,
and disable it on removal.

Change-Id: Iee131f0dca5cdcad573af68a30223a48b22f4cd6
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/94604
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Singh, Murari KumarX <murari.kumarx.singh@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>