[Telephony/MCD] REVERT_ME: Add wake lock during modem reset
BZ: 90692

As GPIO used for RESET_OUT signal on ctp_pr2 can't wake up the platform, and
as the platform can go to sleep while a reset is ongoing, we should get a
wakelock to be sure to receive the IRQ. This is a workaround, useless on
CTP_PR3 as the GPIO is then a gpio_aon one.

This patch can be reverted once the wakelock is managed by MMgr. See BZ92428.
This patch can also be reverted once CTP_PR2 hardware are not used anymore.

Change-Id: Ibf66440e4ecd462cbbb8b7e02450403afe7233a3
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>