audio: fix the rtpm count issue with power control dai
BZ: 90966

Currently the sst power up and down calls are done in hw_params
and hw_free callbacks for Power control dai. In this case if the
power_up_sst fails due to intel_sst_check_device error, then the
hw_free is called by asoc core as part of error handling, which
calls power_down_sst. And then the client which opened the power
dai also closes the device, triggering the power_down_sst. Thus
rtpm count goes negative.

Fix is to do the sst power up and down in the the startup and
shutdown callbacks for power dai.

Change-Id: Ia4cec6dbbad28a32609ac6bc195e0fe759075632
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Reviewed-on: http://android.intel.com:8080/95186
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Kalliguddi, Hema <hema.kalliguddi@intel.com>
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>