[REVERTME] modpost: treat section mismatch as error
BZ: 91318

Despite the effort to remove section mismatches from our kernel, we
always end up with more. This patch forces section mismatch warnings to
be threated as error to block new issues from being integrated.

Since developers on community are already discussing a proper way to
implement such support, this patch should be removed as soon as we've a
solution to backport.

Change-Id: I89ff6ca2024c72b3fb271ecec0f31bf969d11209
Signed-off-by: David Cohen <david.a.cohen@intel.com>
Reviewed-on: http://android.intel.com:8080/95435
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>