mmc: enable eMMC4.5 host controller for Baytrail B0
BZ: 92148

Baytrail B0 has two eMMC host controllers. One is called 4.41
and the other one is called 4.5.

This patch is used to enable the 4.5 host controller, as well as
the additional HS200 feature for the eMMC devices which can support
this. But we will first to disable HS200 due to the eMMC little
rock board needs a rework to support HS200, otherwise failure will
be encountered

Change-Id: I509a2f2323edd82556f987477e3e71783dc99a13
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/96279
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>