Display: [port from K3.0] correct mipi overlay flip legal check
BZ: 92678

This patch is to fix the green display on MIPI when plug in HDMI while
playing video during a incoming call. It was supposed that when HDMI
is connected, overlay should always be switched to pipe b for extend
video mode. But it is not true under the incoming call case. So need
to correct the flip legal check.

Change-Id: I8770a15a9e6e17c5927cef70eb728a963f694c14
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/96499
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Xie, Lin <lin.xie@intel.com>
Reviewed-by: Liu, Hong <hong.liu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>