Graphics: Remove the DRM LOCK for graphics kernel driver
BZ: 81233

In graphics kernel driver, there is already PVRSRVLock in the ioctl
functions, so the DRM LOCK is not needed for graphics kernel driver.

If the DRM LOCK is used, it will cause the mutex starvation, one
graphics thread may not get the GPU time for a long time, the UI would
get stuck.

Change-Id: Ic637cd2ef03b0c2a5b87a57c39697f9d2bdd4ecb
Signed-off-by: gwang23 <gang.a.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/97868
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Xu, Huiying <huiying.xu@intel.com>
Reviewed-by: Dai, Yu <yu.dai@intel.com>
Tested-by: Dai, Yu <yu.dai@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>