imx135: do not allocate unused VPROG1 and GPIO 176 for Victoriabay
BZ: 95542

imx135 platform code allocates VPROG1 and sets it to 2.8 V
on all platforms, including Victoriabay. Victoriabay main
camera does not use VPROG1 which is connected to 2nd camera
and should be set to 1.2 V. Setting the voltage to 2.8 V
may damage the secondary camera.

Check for VB and do not use VPROG1 nor GPIO 176 on VB.

Change-Id: Id80a753eb8eb3940d424cb7af97f792cfa55c698
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Reviewed-on: http://android.intel.com:8080/98222
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Friis, Soren <soren.friis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>