Removed the work-around for the MTX_CMDID_ISSUEBUFF's interrupt enabling flag taking no effect bug.
BZ: 95698

Issue: TopazHP's JPEG firmware had a bug and there
was always an interrupt when JPEG encoding's
MTX_CMDID_ISSUEBUFF was completed, although we had
cleared its interrupt enabling flag. So a work-around
was made to ignore unexpected MTX_CMDID_ISSUEBUFF's
interrupts. And now the firmware was just upgraded
and that bug was fixed.

Solution: So we could remove that work-around.

Change-Id: Idb6486d6c5096fd135e8c1c52bf3546de77c2dcd
Signed-off-by: SUN,Jing <jing.a.sun@intel.com>
Reviewed-on: http://android.intel.com:8080/98327
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lin, Edward <edward.lin@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>