[REVERTME]sdhci: enable mmc irq before suspend for Merrifield
BZ: 78171

On currently mainline, for some unkown reason, the Merrifield
eMMC0 host controller (and some other components) irq slot entry
in IOAPIC RTE is masked before the eMMC driver suspend callback
completes. Since eMMC driver needs interrupt to send CMDs to card
during the suspend callback, then it's an issue and finally
blocked the S3 enabling on Merrifield platform.

Before the really root cause identified, just submit this patch
as a workaround. This patch just calling enable_irq at the
beginning of the sdhci driver supsend callback and disalbe_irq
at the end of the same function.

This is just a workaround patch to un-block the S3 enabling work
on Merrifield platform. Once the really root cause identified
and fixed, this patch needs to be reverted immediately.

Big thanks to Zayet, Aymen for contributions of this patch.

Change-Id: I4199d8b5c25230acd282c23ff0740bc85054e187
Signed-off-by: Yunpeng Gao <yunpeng.gao@intel.com>
Reviewed-on: http://android.intel.com:8080/98738
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Dong, Chuanxiao <chuanxiao.dong@intel.com>
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>