[PORT FROM R42L] storage: Don't switch VCCSDIO when SDIO default off in VFLAGS
BZ: 96944

Use VFLAGS to turn off VCCSDIO as SDIO0 no
used, In some projects, VCCSDIO should be dynamic controlled when
suspend/resume. So we can consider SDIO0 no used when VCCSDIO is
turned off by VFLAGS, VCCSDIO switch on/off is not allowed in SDHCI
at this condition. When we set VFLAGS to turn on VCCSDIO as default,
the sdhci driver allows pm to control VCCSDIO dynamically for
saving power.
This function dependent on platform, so merge the platform relative
code.
Now VFLAGS is not set, so this patch can't be tested.

Change-Id: I75c6bbdfee70ef8e7e9e25e3c5fe90130ffab48b
Signed-off-by: Ning Feiyi <feiyix.ning@intel.com>
Reviewed-on: http://android.intel.com:8080/99363
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>