[PORT FROM JB42L][Telephony/MCD] Don't handle Core Dump INT when modem reset
BZ: 95809

when Modem is in reset procedure, core dump INT is expected
to raise but it shall be seen as "spurious" and ignored --
This is what FFL driver did for 6260. For 6360, modem control
driver shall also obey the same rule, or else, if MCD handle
Core Dump INT when modem reset is ongoing, it will hit failure
when read core dump file, and then initiate unnecessary modem
reset, this will make modem disorder.

Change-Id: Ic5865c7ac521a13c6e52856a1e2da77352b99769
Signed-off-by: chao.bi <chao.bi@intel.com>
Reviewed-on: http://android.intel.com:8080/99471
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>