mfd: check driver data on suspend.
BZ: 97840

During suspend on Tangier based platform, the suspend crashed
with "unable to handle paging request issue". This is because
the suspend callback is trying to suspend the 0000:00:04.0 hsu
global registers(useless for pnw/clv/tng),so pci_get_drvdata(pdev)
is returning a null pointer.

To avoid the issue, drvdata point is checked and suspend/resume
callbacks returns immediately if device is the hsu global registers.

Change-Id: I14b84241c74ffe6ddf3f11f3e763c7d49e3ef8db
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Signed-off-by: Wang Yiyang <yiyang.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/100116
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>