[PORT FROM 4.2]max17042: add fake_batt_full option.
BZ: 48314

Some platforms like pnp boards report a low capacity
even when they are connected full time to
a charger and they dont have a battery.

Android will force a reboot when it detects a low capacity
on the battery.

To make it easier to work with these platforms and
to prevent unnecessary constant reboot add a boot arg option
to fake a full battery.

Orig-Change-Id: Ie43eb8efc4839f1f39dd30280894efeb9d42c404
Orig-Change-Id: I78f056d85d965756141a84acc3ee2e8f84493894
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Signed-off-by: Ramakrishna Pallala <ramakrishna.pallala@intel.com>

Conflicts:

	drivers/power/max17042_battery.c

Change-Id: I6baebc38dc52cc2e078f86869cb45806ca78a0d8
Signed-off-by: Madhukumar <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-on: http://android.intel.com:8080/100205
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>