[REVERTME][PORT FROM LKML]gpiolib-acpi: introduce acpi_get_gpio_by_index() helper
BZ: 98448

*** Backporter's message ***
This patch is currently under review in LKML. Refer to below link for details.
http://lists-archives.com/linux-kernel/27814359-gpiolib-acpi-introduce-acpi_get_gpio_by_index-helper.html

We can't wait for merging of this patch to upstream but have to pick up from
LKML immediately because most of our drivers depend on gpio acpi support.

Please revert this patch and backport the equivalent patch from upstream
kernel once this LKML patch is merged to upstream.

*** Original commit message ***
Instead of open-coding ACPI GPIO resource lookup in each driver, we provide
a helper function analogous to Device Tree version that allows drivers to
specify which GPIO resource they are interested (using an index to the GPIO
resources). The function then finds out the correct resource, translates
the ACPI GPIO number to the corresponding Linux GPIO number and returns
that.

Change-Id: I4e25c34464ae8fae493622c8c37f4cd2960abcc9
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-and-Ported-by: Bin Gao <bin.gao@intel.com>
Reviewed-on: http://android.intel.com:8080/100366
Reviewed-by: Cohen, David A <david.a.cohen@intel.com>
Reviewed-by: Gao, Bin <bin.gao@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>