intel_soc_mrfld: add Wake source definitions for merrifield
BZ: 98949

on Medfield/CTP we have LSS# definitions same for WKC# in the
former case we had two bits defined per LSS and in later one
bit defined, but essentially the bit definitions were one to one.

On Merrifield this one to one definition no longer holds good.

Changing this to the correct definitions as defined in SCU

Change-Id: Ibf93c280369a4059bf70ffcb3e7acf47370a8706
Signed-off-by: Illyas Mansoor <illyas.mansoor@intel.com>
Reviewed-on: http://android.intel.com:8080/100618
Reviewed-by: Ramachandra, Sudarshan N <sudarshan.n.ramachandra@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Rudramuni, Vishwesh M <vishwesh.m.rudramuni@intel.com>
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>