[PORT FROM R42] io_apic: Mechanism to signal kernel panic to SCU
BZ: 98969

IO_apic write generates IPC2 event in SCU. Make a dummy
write to unused location to hit IPC2 event to SCU.
Target of the dummy write is at the end of the RTE table.

SCU can check the value and hit an SCU status dump to SRAM.
This mechanims is always available when kernel is running.
It is not blocked by ipc1 or pmu operations. This is used only
in error situation (panic / warn) to trig SCU to refresh its
status to shared RAM.

Change-Id: I6ae52fad464a47ec334be12c58444ccd9ab7705c
Signed-off-by: Samu Onkalo <samu.onkalo@intel.com>
Signed-off-by: Jouni Hogander <jouni.hogander@intel.com>
Reviewed-on: http://android.intel.com:8080/108464
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>