usb:dwc3-device: Fix back to back setup packets issue
BZ: 97213

The fix is backported from kernel 3.8. In compliance test TD7.06, host
sends back-to-back setup packet, and driver is blocked at waiting for
data XferNotReady event which is going to happen. To fix the issue,
driver won't wait for the event anymore but starts the transfer for data
stage immediately.

Change-Id: I32ab12843e96a9c79a54e77797b8045bb488f1dc
Signed-off-by: jzhuan5 <jin.can.zhuang@intel.com>
Reviewed-on: http://android.intel.com:8080/100669
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Wang, Ran <ran.wang@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>