Revert "[Telephony/MCD] REVERT_ME: Add wake lock during modem reset"
BZ: 90692

As GPIO used for RESET_OUT signal on ctp_pr2 can't wake up the platform, and
as the platform can go to sleep while a reset is ongoing, we should get a
wakelock to be sure to receive the IRQ. This is a workaround, useless on
CTP_PR3 as the GPIO is then a gpio_aon one.

This patch can be reverted once the wakelock is managed by MMgr. See BZ92428.
This patch can also be reverted once CTP_PR2 hardware are not used anymore.

This reverts commit a67fd89d0f71379324ce9b104e015e634ee9e958.

Orig-Change-Id: Ibf66440e4ecd462cbbb8b7e02450403afe7233a3
Change-Id: I9b0db7859c0a36e45f8e9e8935e564b87443ccb5
Signed-off-by: Frederic Berat <fredericx.berat@intel.com>
Reviewed-on: http://android.intel.com:8080/101260
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>