i2c: enable/disable i2c controller properly
BZ: 97844

The correct way to disable or enable the controller is to wait until
DW_IC_ENABLE_STATUS register bit matches the bit we program to the
DW_IC_ENABLE register. This procedure is described in the DesignWare I2C
handbook.

This patch is originally from Mika Westerberg <mika.westerberg@linux.intel.com>.

Change-Id: Ifaa2af6eba8c6fb1f59f7650cf54164edc9738c2
Signed-off-by: Li, Ning <ning.li@intel.com>
Reviewed-on: http://android.intel.com:8080/101507
Reviewed-by: Chen, Jason CJ <jason.cj.chen@intel.com>
Reviewed-by: Yang, Bin <bin.yang@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>