crystalcove_pwrsrc: add reset source indication
BZ: 100413

From Crystal Cove PMIC registers.

The PMIC contains three registers which are intended
to store the cause of shutdown, reset and wake events.

Read and decode these registers at module init so that
the cause of the previous shutdown/reset/wake events
can be found in the logs if needed, and then clear the
bits for next time.

Change-Id: I53289282d4ec4581099e031e3cc000d0e197d2f5
Signed-off-by: Olivier Fourdan <olivier.fourdan@intel.com>
Reviewed-on: http://android.intel.com:8080/101814
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-by: Kundella, SreedharX <sreedharx.kundella@intel.com>
Reviewed-by: Mitra, DebalinaX <debalinax.mitra@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>