[PORT FROM R42-STABLE]Audio: Fix multiple reporting of jack for spurious interrupts
BZ: 90196

With the current implementation every HS insert/remove event is reported
even for the spurious interrupts.

This causes the the route manager change the policy from IHF to headset
and vice versa for every event reported.

With this fix the jack event is reported only after insert state is
confirmed and retries are done to identify between a headphone and mic.

Also changed the jack detection irq_flags to use both rising and falling
edge trigger instead of only falling edge.

Change-Id: I1b7913a06e0ea208b8e33119177f646d946e1ea5
Orig-Change-Id: I2328fddf1748dcd03e8deb4845158a4bb7b8778d
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Reviewed-on: http://android.intel.com:8080/102553
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Joshi, SantoshX <santoshx.joshi@intel.com>
Tested-by: Joshi, SantoshX <santoshx.joshi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>