[PORT FROM R42S] [Telephony/eDLP_Driver]: return err code when dlp probe fail
BZ: 101969

When DLP driver fails to probe, it still return 0 per current code,
then the base driver would understand that DLP driver is installed
successfully, and will continue to bond dlp driver to devices and
it's risk to lead to wrong status.

This patch is to return error code in dlp_driver_probe() when it
fails.

Change-Id: I0341d8305b29fb75e8aa4bbbf8aa3bb218ee3484
Orig-Change-Id: I9dd93f358338ec3d3022c4ce45063b5c138957f7
Signed-off-by: chao.bi <chao.bi@intel.com>
Reviewed-on: http://android.intel.com:8080/110015
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>