[REVERTME] Don't enumerate ACPI devices using I2C4 bus
BZ: 102071

This is because camera handles i2c registration based on atomisp2
platform data, and gets registered in a controlled sequence.

Currently camera still uses fake SFI table for camera sensor/flash
registeration, and not implement ACPI machanism yet.

This is a temperarary solution to disable ACPI for I2C4 bus, which
only have camera devices connected. When design of ACPI for camera
finnished, we should remove this.

Change-Id: I6eb4e8a1ffb8038d70646544cc458e26717cba33
Signed-off-by: Sergio Aguirre <sergio.a.aguirre.rodriguez@intel.com>
Signed-off-by: Tao, Jing <jing.tao@intel.com>
Reviewed-on: http://android.intel.com:8080/102747
Reviewed-by: Huang, Shenbo <shenbo.huang@intel.com>
Reviewed-by: Yang, Bin <bin.yang@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>