[PORT FROM R42S] [Telephony/eDLP_Driver]: set flow_stopped when dlp tty close
BZ: 101893

When dlp tty close, it's possible that TX waiting FIFO is not
empty. Close will be blocked in tty_wait_until_sent because of
buffered.

The patch is to set flow_stoped when dlp tty close, let tty frame
to recycle the whole content of the TX waiting FIFO before
tty_wait_until_sent.

Change-Id: I8b8c0e8637bdcf84fd986c7608d676a79670ae12
Orig-Change-Id: I9116b991a1b5163bed333c8890b7c9f04afcafc1
Signed-off-by: xiao jin <jin.xiao@intel.com>
Signed-off-by: zou juan <juan.zou@intel.com>
Reviewed-on: http://android.intel.com:8080/104698
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Tenoutit, FaouazX <faouazx.tenoutit@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>