VLV2: Disable the legacy timer
BZ: 104490

The bug found a weired thing, that enabling VLV2 HSU driver with mfd.c
will cause the irq0 (legacy timer) to increment very quick, as seeing
millions in several minutes after boot.

The root cause maybe some irq routing problem, so that the FASTIRQ 0
is triggered by mistake somehow, usually the irq 0 is allocated for the
legacy PIT timer.

This patch will disable the legacy timer and irq0, as it is not needed
since OS is relying on AON LAPIC timer, just like the MFLD/CLVP/MRFLD
cases.

And better, we may still need to figure out why IRQ0 is triggered, from
both the FW and OS side.

Change-Id: I3d29b9f81d5df1f3c20e8beb4c6e77c8baddb9a7
Signed-off-by: Feng Tang <feng.tang@intel.com>
Reviewed-on: http://android.intel.com:8080/105293
Reviewed-by: Marotte, Fabien <fabien.marotte@intel.com>
Tested-by: Marotte, Fabien <fabien.marotte@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>