EM/THERMAL: Enable BYT thermal driver interrupt handling
BZ: 100228

This patch enables BYT thermal interrupt handling.
Through the sysfs interfaces the thresholds can be
configured, and the PMIC generates interrupt when
the thresholds are reached. This interrupt is
received by the PMIC driver, and forwarded to the
thermal driver, after first level processing.

This patch also updates the GPADC driver to
return the 'exact' return (instead -EINVAL)
when GPADC sampling fails.

Change-Id: Id0d45a820c7d3f3565b132122551a3520952bbf3
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Reviewed-on: http://android.intel.com:8080/106214
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>