audio: fix dma max block size for mrfld
BZ: 106311

FW download was failing when the dma block size exceeded the max
dma block size. In such case the dma transfer needs to be split
into two blocks.

Fix is that the max block size should be set to (2*17 - 1) instead
of (2*17) because block transfer size is valid for 17-bits only.

Change-Id: I3248df8b9b5456cd6ad21f985e9e3b2ab30a5f68
Signed-off-by: Lakshmi N Vinnakota <lakshmi.n.vinnakota@intel.com>
Reviewed-on: http://android.intel.com:8080/106554
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>