audio: Enable Third DMIC for Victoriabay variant
BZ: 95329

This patch provides a mixer control to select the DMIC1 or DMIC3
for the Victoria Bay varient.

There is a switch (component U1AU10) for selecting the Mic 1 or
Mic 3 to the codec right channel. The switch is controlled by SoC
GPIO pin GP_AON_088. The control signal name is DMIC_1_EN.
This patch writes 1 to DMIC_1_EN when the Mic 1 is selected
and 0 when Mic 3 is activated.

Change-Id: Id8b66bbb7a9c009948aafc05fe230f6b867fea79
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/102901
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Joshi, SantoshX <santoshx.joshi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>