Fix UART mux, debug port not working
BZ: 105076

This fixes the uart mux code trying to erroneously uninitialize the shared
port before it is initialized.

In addition, serial_port_setup will eventually cause a call to
serial_hsu_console_setup, if the port is console. Therefore, copying
uart_hsu_port struct to the alt_up has to be done before calling
serial_port_setup. Otherwise the console setup will be copied to the shared
port as well.

Change-Id: I814cf44f53065de67b955f9c8d3f74dde943c1a1
Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>
Signed-off-by: Wang Yiyang <yiyang.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/105775
Reviewed-by: Du, Alek <alek.du@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>