[Audio] Remove compilation warning in Audience kernel driver
BZ: 107583

The Audience driver a1026.c compiles but produces a warning while
it should compile without any warning.

This patch updates it to avoid warning.

The issue is the use of the i2c msg struct in which the data pointer
is not const, while the write function of the driver receives data
in a const pointer.

In order to fix it, it replaces the use of the i2c msg and the
function i2c_transfer in both read and write functions, and use
the functions i2c_master_recv and i2c_master_send instead.

This avoids the pointer warning since i2c_master_send needs a
const pointer.

Change-Id: I4c94301872ec4c5de7e2dda16942b65f6f96643b
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/107480
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-by: Gross, Mark <mark.gross@intel.com>
Reviewed-by: Lee, JhinX <jhinx.lee@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>