[PORT FROM R42]Audio: [WORKAROUND]Fix for garbled audio after 6 hours
BZ: 93224

Driver reports Delay and the buffer position in Samples to ALSA and these are
u32 counter. The problem is when ringbuffer counter wrap happens, ringbuffer
counter reported is not at the correct offset. This will lead to wrong offset
being reported and the next write will start from a wrong offset,
which can cause garbled Audio.

This patch uses u64 counter to prevent the wrap issue instead of u32.

Change-Id: Ie2b3e00a5ecfc98745336f7d911844fd419d981b
Signed-off-by: hevendra R <hevendrax.raja.reddy@intel.com>
Signed-off-by: Santosh A Joshi <santoshx.joshi@intel.com>
Reviewed-on: http://android.intel.com:8080/110113
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>