toggle CLK pin to abort I2X xfer
BZ: 111302

This patch is used to abort transferring while slave device
is pending on clock waiting.
When the master detects the SDA line stuck in the low state, it
merely needs to send some additional clocks and generate
a STOP condition.
The procedure is as follows:
1) Master tries to assert a Logic 1 on the SDA line
2) Master still sees a Logic 0 and then generates a clock
pulse on SCL (1-0-1 transition)
3) Master examines SDA. If SDA = 0, go to Step 2; if
SDA = 1, go to Step 4
4) Generate a STOP condition

Change-Id: I36b16092d1f8265664b780c49c917f05c97aff3c
Signed-off-by: Bin Yang <bin.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/110098
Reviewed-by: Du, Alek <alek.du@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>