drm/i915: Fix sprite clamp check
BZ: 111329

The clamp check fails when sprite is in the corner.
Sprite plane can move out of primary plane as long as there is
some overlap. If it totally move out there can be hardware hangs.
This patch allows sprite to move out of primary plane slightly but
not totally.

Change-Id: Ie15a86eeabdbaf06b15f5bc4c3415444bf51ae6b
Signed-off-by: Oleg Muraveynyk <oleg.muraveynyk@intel.com>
Signed-off-by: Gajanan Bhat <gajanan.bhat@intel.com>
Reviewed-on: http://android.intel.com:8080/110083
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>