dwc3:gadget: set maxpacket to default value when ep is disabled:
BZ: 108778

Need to reset the maxpacket value of endpoint when it is disabled.
As gadget driver will config the maxpacket value of endpoint, but
the value is not reset to default value after ep released by gadget
driver. This will cause wrong maxpacket issue if the same endpoint
assigned to another gadget function driver.

Change-Id: I09a6e2da0317f8dae19a1ce966077beed855c869
Signed-off-by: Wu, Hao <hao.wu@intel.com>
Reviewed-on: http://android.intel.com:8080/110484
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Zhuang, Jin Can <jin.can.zhuang@intel.com>
Reviewed-by: Luo, Zhenlin <zhenlin.luo@intel.com>
Reviewed-by: Li, Jiebing <jiebing.li@intel.com>
Reviewed-by: Wang, Yu Y <yu.y.wang@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Tested-by: Wang, Ran <ran.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>