audio: populate SST platform data through PCI quirks
BZ: 112883

This patch uses PCI quirks to populate SST platform data.

The SST driver platform configurations such as:
1. information required to load firmware,
2. maximum stream.
3. maximum dma length.
4. number of probe stream supported.
5. Board and platform related configurations that are required to
   dynamically configure the firmware
are all passed through the platform data of SST driver.

The dynamic configuration of firmware was done through the "sst_platform
board file" to the sst driver in the earlier code. Since the PCI quirks
can interface to the SST driver directly, added few clean ups to the
sst_platform and driver interface to remove the older code.

Change-Id: Ie431bfe2bf30afd989dda8f09f44e6f4b83c1e2f
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/109846
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Raja Reddy, HevendraX <hevendrax.raja.reddy@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>