EM/Charger: Fix for battery status not showing FULL
BZ: 108116

This patch fixes the battery status not showing FULL issue. Battery
status is decided by the charger framework once getting hw termination
notification from the driver.
BQ24192 charger depends upon the charger interrupt to notify the hw term
event to the framework. The charger irq gpio was not mapped and driver was
not able to notify the framework.
Now the irq gpio is mapped and charger gets a notification on hw termination
and notifies the framework.

Signed-off-by: Pandey Raj <raj.pandey@intel.com>

Change-Id: Ia1cef4febf7aaecd07bb97580b3a88279c0ab275
Reviewed-on: http://android.intel.com:8080/111235
Reviewed-by: Pandey, Raj <raj.pandey@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Kannappan, R <r.kannappan@intel.com>
Reviewed-by: Kundella, SreedharX <sreedharx.kundella@intel.com>
Reviewed-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>