force to handle i2c IRQ on PNW
BZ: 97844

On PNW, I2C IRQ might come to IA after suspend due to huge latency
of IRQ forward. To support share irq on BYT, it needs to ignore the
IRQ after controller disabling/suspend. It might cause lots of unhandled
irq issue which causes kernel to disable this IRQ.

This patch add new flag to identify the share irq request on BYT. On
BYT, we don't have big latency of IRQ forward. So this patch should be safe.

Change-Id: I82a7c0f72ad1c49d0e1f597332fda11ffc94740f
Signed-off-by: Bin Yang <bin.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/111252
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>