[BT]bcm_bt_lpm: do not return -EBUSY if the chip is OFF during suspend.
BZ: 113126

On baytrail, a pull-down is missing on the gpio bt_host_wake. This one
is used in order to wake up the ValleyView from low power state (S0ix and S3).
Due to that mess, the gpio is not driven when the chip is off, and the
suspend callback is rejected with -EBUSY since the gpio is not pulled down.

From the software point of view, we can make the driver more fault-tolerant
by adding a check in the suspend callback: if the BT device is off then
return 0. Same logic while resuming.

Even if this patch avoid SW issue, the hardware should be fixed and the GPIO
host_wake must be pulled down.

Change-Id: I837dacd262f3e0c4e741d9968a94318101c9e91d
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>
Reviewed-on: http://android.intel.com:8080/111411
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Combes, Patrick <patrick.combes@intel.com>
Reviewed-by: Vamour, Cedric <cedric.vamour@intel.com>
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>