[PORT FROM R42S][Telephony/HSI_controller]: Wait for end of ISR tasklet before flushing RX queues
BZ: 91611

This modification is need when HSI.h interface is used only:
Because the HSI_timeout function is not executed in atomic way
(breach when the hsi_transfer function is called) it exists a
race condition when the hsi controller driver is closed and
reopened while the ISR_tasklet is running. To avoid this
situation, the controller flush function wait for the end
of the ISR_tasklet execution before flushing the RX queues.
This ensures the close of the controller is complete before
being reopened if needed.
In order to ensure the pm_suspend can't happen when there
are pending interrupt status to be handled, a check is done
before going in suspend mode. This must ensure the ISR and
ISR_tasklet are managed atomically from the power management
point of view.

Change-Id: I84343b823ce188b77dceabd026ba7b0f1b561695
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/113308
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>