usb/gadget: enable super speed support for Accessory function
BZ: 112935

The current super speed spoort in Accessory and Audio gadget is not
correct, which leads to ADB not functional if user tries to setup USB3
connection and enable Audio/Accessory function by chance.

This patch provides a fix for ADB issue and enables super speed
connection for Audio/Accessory function.

Change-Id: Iaaed6cbc8a83edc5b51b8e5395add30c686c0de3
Signed-off-by: Jiebing Li <jiebing.li@intel.com>
Reviewed-on: http://android.intel.com:8080/111567
Tested-by: Wang, Ran <ran.wang@intel.com>
Reviewed-by: Wang, Yu Y <yu.y.wang@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Griffiths, Richard A <richard.a.griffiths@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>