EM/Charger: Fix for VB PR1 not charging with digital battery changes
BZ: 113662

Charger driver uses the battery profile structure with capacity as 4 bytes.
The EPROM content for VB PR1 devices has capacity value populated as 2 bytes,
when driver tries to access the capacity, bytes following capacity are getting
corrupted.
This patch fixes the issue by reading only 2 bytes for VB PR1 which is not
corrupting the following bytes and battery profiles are getting populated
correctly.
Added the provision for fixed battery as well. If the FPO[1] is set, driver
will read the battery profile from the SMIP instead of EPROM

Also added the code to dump the battery charging profile to satisfy the CAP-158
requirements

Change-Id: I10ca0f69e2c8c9848e43d5988c7b08e0af95b4aa
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/111714
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Kannappan, R <r.kannappan@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Tested-by: Leveque, Olivier <olivier.leveque@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>