EM/THERMAL: Add thermal sensor list for bodegabay
BZ: 112473

This patch adds the thermal sensor list for bodegabay
to add correlation constants corresponding to PRh
for correct skin temperature prediction.

In this patch the same thermistors and correlation
constants are used from CTP as PRh has the same IO
and body. This initial set of values will be further
updated.

Change-Id: I3459eeec11cf7cde9bff3692965d5c0ea34088bd
Signed-off-by: Ajay <ajay.thomas.david.rajamanickam@intel.com>
Reviewed-on: http://android.intel.com:8080/111802
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Roychowdhury, Saikat <saikat.roychowdhury@intel.com>
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: Singh, NavjotX <navjotx.singh@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>