drm/i915: Fixing the EDID parsing for Vendor specific data block
BZ: 115571

A Vendor Specific data block contains atleast 5 bytes. The size of the data
block is represented by the bits 4..0 of initial byte of the block.
Currently, while EDID parsing it was assumed that the VSDB contains more than
5 bytes and no check was made for the size of the block. This may resulted in
wrong data being parsed as part of ELD.
Adding the check to parse VSDB only if block size is more than 5.

Change-Id: Ida5e6f68af683121aabef11256dec6dd097b98f3
Signed-off-by: Sumit Jain <sumit.jain@intel.com>
Reviewed-on: http://android.intel.com:8080/113103
Reviewed-by: Pandey, Alok <alok.pandey@intel.com>
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>