EM/Charger: Fix for VINDPM settings during low battery boot
BZ: 110075

Low battery boot with DCP charger was seeing a drop in VBUS and the
VBUS oscillation wouldn't allow enough current for the platform to
boot and charge the battery and hence the platform would shutdown.
This fix would keep the VINDPM setting for charger to default if the
vbat is less than or equal to 3600mV. Also VINDPM would be changed
dynamically depending upon the vbat.

Change-Id: I4f965a6daaae19ece0a367af2730f4a0630859bb
Signed-off-by: Pandey Raj <raj.pandey@intel.com>
Reviewed-on: http://android.intel.com:8080/113534
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Reviewed-by: Mitra, DebalinaX <debalinax.mitra@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>