EM/Charger: Fix for battery dead issue on MTBF benches
BZ: 109593

While running the MTBF suites on VB devices, device after reaching
FULL would discharge all the way down to zero % even after enabling
the charging. Charger driver cancels the task worker and once WDT expires
bq charger goes to default values which keeps the inlimit programmed to 100mA
in which case platform starts consuming the battery.
The reason is because once Framework enables the charging for the maintenance
charging, it sends only CC/CV commands to set the charging current and charging
voltage but since inlimit is programmed with 100mA, battery drains.
This patch fixes the issue by not killing the worker and also ensures that
right in-limit is programmed so that platform consumes from VSYS as long as
charger is connected as an when needed.

Change-Id: Id43c9fac4734178048dae6ca0b9bc1f48137d43b
Signed-off-by: Pandey Raj <raj.pandey@intel.com>