drm/i915: Set Punit offset register 0x6 correctly for turbo
BZ: 116249

We are currently setting the entire 0x6 register of punit to zero for
turbo to kick in and for the punit to honour our freq requests. This is
incorrect behavior however. With this patch, we are setting only the 7th
bit of that register to zero. This is a reserved bit but is necessary
for turbo to kick in properly.

Change-Id: I84b9c82ca4b34b023fa9d58fdf8abd4beff05e1d
Signed-off-by: Rohit Jain <rohit.jain@intel.com>
Reviewed-on: http://android.intel.com:8080/113584
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Purushothaman, Vijay A <vijay.a.purushothaman@intel.com>
Reviewed-by: Chevalier, Richard <richard.chevalier@intel.com>
Tested-by: Chevalier, Richard <richard.chevalier@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>