revert [REVERTME] cpufreq: reducing the turbo ratio by two
BZ: 115564

Enabling turbo in punit is causing boot to fail. it has been
rootcaused by punit team & they suggested to reduce the max
turboratio by 2. This seem to help boot all the way up to gui.
This is only a WR until punit team figure out why it is failing
when the limit is set to high. This setting of msr should have
beendone ideally by IA FW , since they said that in windows its
donethis way in bios.

Change-Id: Ie67ed02ae03ea1c6ef9137d37eea666a92e24e78
Signed-off-by: Vishwesh M Rudramuni <vishwesh.m.rudramuni@intel.com>
Signed-off-by: Fabien Marotte <fabien.marotte@intel.com>
Reviewed-on: http://android.intel.com:8080/113673
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>