drm/i915: Adding HW Native sync support
BZ: 133194

This patch requires BZ: 133199 for the matching DRM API changes.

Added support for native sync for batch buffers. Each ring has a timeline
that tracks sequence numbers. When a caller submits a batch buffer they
can request that a fence is returned. The fence will be signalled when the
batch buffer comples or if TDR is triggered.

At present the KMD cannot accept incoming fences. This feature will
require much of the scheduler functionality so for the moment incoming
fences are handled in user mode.

User interrupts have been permanently enabled so that the timeline can
be signalled after every batch buffer completion in case anyone is
waiting on a fence.

The DRM_IOCTL_I915_GEM_EXECBUFFER2 IOCTL has been modified to have
read/write access so it can return a file descriptor to user space.

Added I915_EXEC_WAIT_FENCE and I915_EXEC_REQUEST_FENCE flags that
can be passed in the flags field of an exec IOCTL.

Added a write of the seqno to the hardware status page before each
batchbuffer dispatch. It is cleared immediately after the batch buffer
completes. TDR uses this to detect which batch buffer has hung, i.e.
if it sees a non-zero seqno in the HWS page then it knows that the batch
buffer didn't complete and it can mark the sync fence as timed out.
The GPU must be idled whenever the seqno wraps as the timelines do not
cope with wrapping.

Change-Id: I79519a159144b1d33dd57a2bf74ce8ee841493b7
Signed-off-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Signed-off-by: ian-lister <ian.lister@intel.com>
Reviewed-on: http://android.intel.com:8080/127454
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>