BYT: do not dump full logs on RS232 console upon panic
BZ: 118415

In case of panic, full logs are printed on RS232 console and it takes about 2mn.
if watchdog is enabled, IA is resetted after about 35s, so there is not enought time to write log on EMMC, due to console.

Workaround is to dump only first 2048 bytes to console, and then only write logs to EMMC.

Change-Id: I802cdba96dfe373e48fc07fcb5715187cb73b3e2
Signed-off-by: Laurent JOSSE <laurent.josse@intel.com>
Reviewed-on: http://android.intel.com:8080/115061
Reviewed-by: Berthier, Emmanuel <emmanuel.berthier@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>