[PORT FROM R42S] Fix for SEP driver creates incorrect DMA MLLI tables in a corner case
BZ: 117886

The cause of this bug is a corner case where an intermediate input MLLI table
(i.e., a table that is not the last table) total data size is not a multiple of
the operation block size (which is illegal for the Chaabi DMA). This illegal
situation is caused when calculating the maximum data size of the input and
output MLLI tables, there are more input MLLI tables than output MLLI tables.
Usually the total data sizes of the input and output MLLI tables are normalized
by changing the table total data size to the greatest multiple of the operation
block size that is not greater than the total data size. If the input data size
is greater than what a single MLLI table can hold but the output data size is
less than or equal to what a single MLLI table can hold (i.e., the calculation
produces more input MLLI tables than output MLLI tables), then the MLLI table
data size normalization does not take place (i.e., the input and output MLLI
table total data sizes are normalized only if the output MLLI table is not the
last output MLLI table). This will cause an intermediate input MLLI table total
data size to be a non-multiple of the operation block size, which is illegal for
the Chaabi DMA. The CRYS operation in Chaabi then produces a data input size
illegal error (0xF0000A).

This bug was fixed by checking the last table flag after the input MLLI table
maximum size is calculated. If the input MLLI table is not the last input MLLI
table then its total data size is normalized. This will make sure that the total
data size of both the intermediate input and output MLLI tables are properly
normalized to the largest multiple of the operation block size. Only the last
MLLI table can have a total data size that is a non-multiple of the operation
block size.

Change-Id: I92f0c04d1456a83693d307f6030dcca4a91c3814
Orig-Change-Id: I61a11035d45535798954e3896049bedc8cecbd89
Signed-off-by: Michael Matthews <michael.l.matthews@intel.com>
Reviewed-on: http://android.intel.com:8080/115223
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Pitchai, Loyola <loyola.pitchai@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Shuster, Norvin <norvin.shuster@intel.com>
Tested-by: Shuster, Norvin <norvin.shuster@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>