EM/pmic_ccsm: Release wake lock for USB host mode
BZ: 119049

Release wake lock for USB host mode. Previous implementation had
wake lock acquisition for both vbus and id detection. Now we check
whether the VBUS enabling is happening for a device. If so, we
release the wakelock so that the system can go to S3 if it is in
host mode. For ACA, the system will hold the wake lock.

Change-Id: Ib19621fb4514aa27c9f23c2dcb536a9ac7c06242
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>
Reviewed-on: http://android.intel.com:8080/115553
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>