Audio: free jack gpio upon card registration failure
BZ: 115619

On card registration failure, the jack device is freed but the gpio
resources are not freed. Inserting a HS will cause the interrupt handler
to be called and cause kernel panic during jack report.

On the card instantiation failure status is propaged to the machine
driver through "instantiated" member of the card structure. Check the
card instantiated variable to free up the jack gpio and irq on card
registration failure.

Change-Id: I0ac3c24f493a6211f1042e816cbc2ae6cde5667a
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Reviewed-on: http://android.intel.com:8080/115507
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>