audio: Implementation of IPC message in the shutdown handler.
BZ: 111316

This patch fixes the audio FABRICERR, that was caused during adb reboot,
when the audio playback is active.

An ipc message PREPARE_FOR_SHUTDOWN is implemented in the shutdown handler
to notify the LPE firmware about the shutdown. LPE firmware free's up the
active streams and aborts the active DMA transaction during device shutdown.
This helps to prevent the DMA fabric error during "adb reboot", when the audio
playback is active.

In addition to this, have introduced a new state SST_SHUTDOWN as the shutdown
state for the sst driver, this is just for the debugging purpose to make sure
that no other calls can reach driver after the shutdown state.

Change-Id: Iabb0fc03991a88fed7ebb04f65c1578e5a660754
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/115555
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>