atomisp2: use hard-coded alignment also for video output
BZ: 120802

In case video viewfinder is larger than main output, user-space may
route the main output to display and viewfinder to encoder. When this
happens, the GFX compatible alignment must be used for video output as
well.

This can be removed once all CSS versions support specifying the minimum
width, and when user-space provides the correct stride information
(which is use-case specific so kernel cannot know).

Change-Id: I47ca74ea6d8f04843939c035342436b726beded7
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/117163
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Reviewed-by: Kaartinen, Jukka O <jukka.o.kaartinen@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Tao, Jing <jing.tao@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>