audio : disable MCLK before changing clk divide ratio
BZ: 100066

When headset is plugged in, the MCLK is put in low power mode
during suspend. In this case, the CLK divide ratio is being
changed while the MCLK is on.

Fix is to disable the MCLK before changing the clk divide ratio
and reenable after change. Further to ensure that codec bias level
is in STANDBY, the soc_suspend is called before the mclk switch
is called.

Change-Id: I3aa4074a4746829ff8806903954d1609e5d31555
Signed-off-by: Lakshmi N Vinnakota <lakshmi.n.vinnakota@intel.com>
Reviewed-on: http://android.intel.com:8080/102658
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>