EM/pmic_ccsm: Don't take wake lock for USBID
BZ: 123289

Don't take wakelock for USBID change. The driver need
to take wake lock only if VBUS interrupt is asseted. Currently
the driver is taking wakelock for any USBID change and releases
it when it detects the VBUS change on enabling the VBUS.
But sometimes PMIC reports VBUS ID removal on boot eventhough the USBID
was not connected/removed. This results in acquiring wakelock. This patch
removes the wake lock if the interrupt is for USBID

Change-Id: Ie525599ce2a07ca80f199624c7d324f31e3cdbaa
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Reviewed-on: http://android.intel.com:8080/118528
Reviewed-by: Iyer, Yegnesh S <yegnesh.s.iyer@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>