audio: always do get_params for compress volume get in mrfld
BZ: 122633

Currently, there is an optimization in driver to return the last set_params
value for any get_volume call from MW instead of doing get_params from FW.

For User volume, FW does not save/restore the lib params. So when
MW does a get_volume, a get_params for the lib needs to be send to FW.

Also, fixed incorrect wake block in case of async error messages.

Change-Id: Ie0368c6810bfe2f46d3a13eb4451364377946fa5
Signed-off-by: Lakshmi N Vinnakota <lakshmi.n.vinnakota@intel.com>
Reviewed-on: http://android.intel.com:8080/119221
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>