WA to prevent the failure when running the Google CTS on Renderscript driver
BZ: 124835

The GPU enabled Renderscript driver was failing the Google CTS
(Compliance Test Suite) which was extensively testing allocations/
de-allocations, communicating with KMD via DRM (with User Ptr/vmap patch).

One of the reasons for the failure was the exorbitant time(> 120 ms) spent in
the de-allocation path of Vmap buffer objects. For the de-allocation
(bo_unreference) of each Vmap buffer object, more than 120 ms was being taken
on average. As a result there was a timeout error(10 seconds) raised during the
CTS test execution.
Observed that the call to MMU notfier un-registration function, made when
destroying vmap objects, was the sole contributor of this huge latency.
As a workaround, have delegated the call to MMU notfier un-registration function,
in the kernel's worker thread(through our private workqueue), so that the
concerned execution path is relieved of this huge overhead.

Change-Id: I03edeb3e22b4ee2b60a43751ca3df91df36411ef
Signed-off-by: akashgoe <akash.goel@intel.com>
Reviewed-on: http://android.intel.com:8080/119798
Reviewed-by: S, Deepak <deepak.s@intel.com>
Reviewed-by: Beckett, Robert <robert.beckett@intel.com>
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Wilma, Pawel <pawel.wilma@intel.com>
Reviewed-by: Kachhi, Naresh Kumar <naresh.kumar.kachhi@intel.com>
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Reviewed-by: Smith, Geoff <geoff.smith@intel.com>
Reviewed-by: Cohen, David A <david.a.cohen@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>