Audio: Fix for SRAM ECC error during firmware download failure
BZ: 126240

When the firmware download fails there is a wait timeout happening in audio
driver, since the FW_INIT message is not received from firmware. Driver dumps the
SRAM exception region during wait timeout and since the firmware is not
alive at this instance, there is an SRAM ECC error.

This patch adds a check for the firmware state and reads the SRAM area only
if the firmware is in the running state.

Change-Id: Id9a7e06bc86266962655eb076c23f3a70f82a5af
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>
Reviewed-on: http://android.intel.com:8080/120670
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Kalliguddi, Hema <hema.kalliguddi@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>