Cleanup shutdown process in intel-mid
BZ: 120649

The goal of this patch is to cleanup some code and to avoid
using power supply functions from OSIP driver when powering
off.

Now the reboot to COS is made depending on a decision taken
at userland (Java).
With the new behavior some uneeded code can be removed now.

Also removed calls to pmu_power_off as this is already
set in intel-mid.c in intel_mid_power_off().

Change-Id: I5995feaadf5aaf5f3e8829431c026d1308210540
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/121800
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>