drm/i915: Increased hang check period and improved TDR detection algorithm.
BZ: 127183

Updated the default hangcheck period to 1000ms. The actual hang
detection time is 3 - 4 times this value, resulting in a minimum
detection time of 3 seconds. The previous value of 250ms caused
some spurious hang detections.

Added a new i915_hangcheck_sample function. This is a combination
of the old i915_hangcheck_elapsed and i915_hangcheck_ring_idle
functions. This allows the idle condition to be checked in
conjunction with the position of the ring pointers so that it
only detects stuck waits when the hardware has been idle
at the same position for 3 consecutive samples. This prevents it
accumulating stuck waits for different batch buffers.

Sample values are now stored for every sample so that on the next
sample we always compare against the previous sample.

Hang check sampling now uses the head pointer AND the active head.
If the same underlying buffer object is used for multiple batch
buffer calls then the active head could appear to hang in the same
place, but the hangs could belong to different do_exec calls.
The recovery works on batch buffer granularity so we need to
distinguish between batch buffers. The head pointer will always
have moved on for a new batch buffer as it requires a new
MI_BATCH_BUFFER_START command in the ring, so by using both head
and active head we can guarantee that hangs sampled for one
batch buffer do not count towards hangs recorded against future
batch buffers.

Added missing DRM_DEBUG_TDR macro for when debug is disabled.

Additional debug added.

Change-Id: Iaa7023ee10180eeb5c9383e19726870278822eaa
Signed-off-by: ian-lister <ian.lister@intel.com>
Reviewed-on: http://android.intel.com:8080/121979
Reviewed-by: Beckett, Robert <robert.beckett@intel.com>
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Ewins, Jon <jon.ewins@intel.com>
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>