drm/i915: Check command ring status in page flip
BZ: 125436

It might be possible that the command ring virtual_start might
be accessed after it is unmapped in a flip. This patch will
add additional check on the ring's status and return EBUSY if
the ring is not accessible in the flip function. Please check
BUG 125436 Comment 6 for the further rational on it.

v2: Return -EINVAL instead of -EBUSY
EBUSY, causes UMD to wait for flip causing system hang. If we return
EINVAL, the reposnse is much better There are some fixes on this in
review and testings. Once they are perfect, we can think of the final
patch

Change-Id: I1158d6cea4bcf8cb543668015044dc7210b9009c
Signed-off-by: Joon Jung <joon.jung@intel.com>
Signed-off-by: Shobhit Kumar <shobhit.kumar@intel.com>
Reviewed-on: http://android.intel.com:8080/122103
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>