atomisp2: create separate folders to compile drivers for different platforms
BZ: 83194

The CSS library used by the atomisp2 driver is specific to the ISP in the
SoC. Even if the API offered by the CSS library is mostly compatible, the
libraries needed on different ISPs is still different.

The atomisp2 driver can use different variants of the CSS library. This is a
compile time choice. This patch create serveral folders to contain the
symbolic links for source code and a specific makefile to compile for the
specific platforms.

Change-Id: Id3dcd1d9da8cfe9d77c6a134fdf52568bf599628
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/122430
Reviewed-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>
Tested-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>