[PMU/MRFLD]: Check for NC and SC device status before entering S3
BZ: 125090

During S3 flow there could be a possibility that S0ix aborts happens.
As seen in BZ 114607. But today due to the lack of s0ix abort feedback
mechanism the platform remains in kind of unknown state until a power
button or USB is connected. This gap is addressed in BZ 12353.

This patch adds a check in PMU driver to check NC and SC device status
before requesting C7 (basically C6 mwait + s0ix hint). S3 flow is aborted
in case the check fails and retries; rather leaving platform in unknown
state.

Change-Id: Id9b676a5f524ad2ff9148baee37bf91c05c0ea83
Signed-off-by: Ananth Krishna <ananth.krishna.r@intel.com>
Reviewed-on: http://android.intel.com:8080/119948
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>