drm/i915: Add check for HDMI encoder is in use for audio usecase
BZ: 127098

This is seen happening when audio is being played and screen blank
is initiated. This patch adds check for HDMI encoder in use for the
audio use case. In the current suspend flow, UMD is also calling
modeset again with null FB to disbale the display causing HDMI to stop
as HDMI pipe is disabled. This should not be done. Now the kernel has a
check for this by checking if HDMI audio registers are in use and does
not allow to disable the hdmi pipe.

Change-Id: Id6ba2f811a61bde294ed0a49e7a96fb12b626880
Signed-off-by: Shobhit Kumar <shobhit.kumar@intel.com>
Signed-off-by: Uma Shankar <uma.shankar@intel.com>
Reviewed-on: http://android.intel.com:8080/123256
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>