[EM/BCU]: Adding support for VWARN1, VWARN2 and VCRIT interrupt count sysfs i/f
BZ: 129333

This provides the VWARN1, VWARN2 and VCRIT interrupt count sysfs interface for
debugging purpose and the same can be used by Kratos to log the data. Also the
interrupt flag IRQF_TRIGGER_RISING is removed, because this interrupt is
directly routed by SCUFW and also added IRQF_NO_SUSPEND. This is required to do
not disable this IRQ during suspend.

Change-Id: Ib527c9295321bab6caef608363ab50f86a3bba1b
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>
Reviewed-on: http://android.intel.com:8080/124357
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>