mei: use a dedicated workqueue instead of the system workqueue
BZ: 117229

All mei tasks are queued on the system workqueue. When the module is
removed or the driver is stopped like at shutdown time, the mei_stop()
function is called and perform a flush of this system workqueue to
ensure all the mei tasks are performed. On rare cases but
unfortunately more likely on android system during shutdown, some
tasks (like filesystem changes) present in this system workqueue
cannot be performed and the system workqueue should not be flushed.

This patch provides a simple solution where the mei driver use its own
workqueue.

Change-Id: I19c593da546f1c45470c93a5e61ec544b7c79114
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: http://android.intel.com:8080/126077
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Usyskin, Alexander <alexander.usyskin@intel.com>
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: Allyn, Mark A <mark.a.allyn@intel.com>
Tested-by: Allyn, Mark A <mark.a.allyn@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>