drm/i915: Update VBT data structures to support generic MIPI
BZ: 142688

Add support for two new blocks #52 (mipi configuration) and and #53
(mipi sequence) in VBT parser

This is the first patch in the series to enable Generic MIPI driver support
The reason to push for r42b-stable is that China CTS team is looking for adoption
in China ecosystem and there are multiple panels to be enabled in a short time
frame. They are looking to base their enabling effort on r42b-stable. This generic
framework will ease this effort.

Still the command line parameter for specific panel as currently exists today will
override this implementation and will continue to work as it is working today. In
that case this generic MIPI driver will not be used at all and static subencoder
drivers will continue to be used. So don't think this will cause impact on oevrall
stability of r42b.

Change-Id: Idcfed27911bff57d96235d903c1f5d4996ec1cb1
Signed-off-by: Shobhit Kumar <shobhit.kumar@intel.com>
Reviewed-on: http://android.intel.com:8080/135981
Reviewed-by: Purushothaman, Vijay A <vijay.a.purushothaman@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>