drm/i915: invalidate tlb for hung ring in reset path
BZ:131012

During MTBF test UIWDT issue is seen. This is rootcaused to HW Status
page not getting updated with the latest seqno, causing ring reset.
Adding TLB invalidate in ring reset path as well to make sure we are
not again reading a stale address from the TLB after this reset.

Change-Id: I218ab1f3fa9f66412b50d4954aabd43491e4c663
Signed-off-by: naresh kumar kachhi <naresh.kumar.kachhi@intel.com>
Reviewed-on: http://android.intel.com:8080/127017
Reviewed-by: Lister, Ian <ian.lister@intel.com>
Reviewed-by: S, Deepak <deepak.s@intel.com>
Reviewed-by: Kamble, Sagar A <sagar.a.kamble@intel.com>
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>