Revert "audio: rt5642 - enable analog power in prepare"
BZ: 133553

This reverts commit 416a6cbc3ef9c2188fbb11ab9ccaf56a5ce9582e.
The original patch moved powering on codec analog rail from bias standby
to bias prepare. This was to save power when moving from suspend to resume.
However, this can cause headset detection issues when HS is inserted when bias
is off. On headset insertion, bias state would get changed to standby, but since
the analog rail is not on, HS would be detected as headphone.

Hence revert this patch. the power saving intended by the original patch will
instead be provided by  anothr patch for BZ 133553

Change-Id: I8ba83cc99d9958da218bd3acb9e515af1c088abe
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>
Reviewed-on: http://android.intel.com:8080/127744
Reviewed-by: Shah, Hardik T <hardik.t.shah@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>