[PORT FROM MAIN]sdhci: Fix possible sleep in atomic context
BZ: 125011

sdhci_enable_sdio_irq() might be called in atomic context when it's called from
mmc_signal_sdio_irq(). This patch removes the call to runtime_pm_get/put from
within this function as sdhci_enable_sdio_irq_nolock() checks if the host is
suspended so it is safe to be called even when the host is in suspend.

Change-Id: I359d3f83593fb367fa18e73d9aa791521946dae6
Signed-off-by: Quentin Casasnovas <quentinx.casasnovas@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-on: http://android.intel.com:8080/127868
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>