EM/THERMAL: Remove workaround to hardcode TjMax
BZ: 136531

The coretemp driver had a workaround to hardcode
TjMax as 90C for all TNG/BYT platforms. This was
added since the TjMax register had a wrong value.

With latest Firmware versions, this is fixed. All
TNG/VLV2 platforms have the right TjMax values.
So, this patch removes the workaround.

Change-Id: I597c8b724b8f3ed0a5abc3f307f370c8994a75de
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Reviewed-on: http://android.intel.com:8080/130401
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>