atomisp2: use mutex to avoid open is called before probe is done
BZ: 138191

bug analysis:

camera hal would open device node to get camera info when
system has booted up. but the probe may not finish at that time,
especially when reserve memory pool. in this bug, the camera FW load
is not done when device node is open. So here comes the error.

solution:
use a mutex to ensure open is run after probe.

Change-Id: Ief9d8ddc9787c53594ba53f617057bfbeadc3c27
Signed-off-by: Jianxu Zheng <jian.xu.zheng@intel.com>