drm/i915: During standby PMIC PWM signal is at 0.39 percent
BZ: 139645

After the system has put to sleep. The backlight is not completely off.
There is still a small pulse observed in the PMIC PWM signal. The issue
is that the PWM duty cycle controls the PWM signal from 0.39% to 100%.
Since android display driver was setting the duty cycle to 0. still 0.39%
of PWM signal was observed. The fix for this issue is that PWM signal has
to disabled in the PMIC PEM CLK DIV register.

Change-Id: I64d9bc8b8a74647ba31f88283fde074e39d1a96a
Signed-off-by: Yogesh Mohan Marimuthu <yogesh.mohan.marimuthu@intel.com>
Reviewed-on: http://android.intel.com:8080/133103
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>