drm/i915: Simple workarounnd for sprite based tearing
BZ: 142179

Sprite plane tearing was being seen especially in low
memory systems as the plane that was onscreen got unpinned
while it was still onscreen. This means that it can be
removed from the GTT while it was onscreen.

This simple fix just holds the object being replaced in a
variable, which then gets unpinned on the next setPlane
call. As the HWC is known to wait at least a frame before
issuing another setPlane call for this plane, this fix will
avoid the flicker.

Change-Id: Iac306b9d82fd734b8a1aebaf257eadb9fe0a8ba8
Signed-off-by: Gary Smith <gary.k.smith@intel.com>
Reviewed-on: http://android.intel.com:8080/136528
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>