drm/i915: fix system lag when quick power off/on with HDMI connected
BZ: 146565

This issue a regression from patch 136066. It is because during quick
resume, hdmi encoder dpms on does not commit and the hdmi port is still
disabled after resume. Then all flip on HDMI will be failed and cause
system lag.

The cause of hdmi dpms-on does not commit is due to encoder_funcs->inuse
incorretly returns true in drm_crtc_helper_set_mode, and then skips
encoder mode set.

This patch adds hdmi port control register check in inuse() to make sure
the encoder status check is correct.

v2:
Check Bit 6 (Audio output enable) in the HDMIB register to validate if
encoder is in use.

Change-Id: I59d7f86baff7b1734cca94be6b8b0f30aa457a0e
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>
Reviewed-on: http://android.intel.com:8080/140219
Reviewed-by: Purushothaman, Vijay A <vijay.a.purushothaman@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>